refactor: hoist remaining nested view/modal components out of App (#41)
Completes the issue #41 audit: the other components defined inside the
LoveTimeApp body all shared the same remount-and-lose-state bug as
ScriptUploadModal — their function identity changed on every App re-render,
so React unmounted + remounted them, discarding local state and re-running
effects.

Hoisted to their own files at module scope, threading dependencies through
props:

- CalendarDatePicker → src/components/CalendarDatePicker.tsx
- ForeplayView       → src/components/ForeplayView.tsx (co-locates its
                       activity/position/combo data)
- GamesView          → src/components/GamesView.tsx (builds romanticGames
                       from customMemoryQuestions/customEmotions props)
- CoinShopView       → src/components/CoinShopView.tsx
- OurJourneyView     → src/components/OurJourneyView.tsx
- AuthModal          → src/components/AuthModal.tsx
- CalendarView       → src/components/CalendarView.tsx (co-locates
                       compressImage + Chinese-date helpers)
- ConflictView       → src/components/ConflictView.tsx (co-locates
                       conflictResolutions)

Exported the shared types they need from App (Notification, ForeplayActivity,
PositionSuggestion, CoinGift, IntimateRecord, JourneyMilestone, AuthState),
consumed via type-only imports so there is no runtime circular dependency.
App.tsx shrinks by ~3,650 lines; behavior and markup are unchanged.
