+++ src/App.tsx
@@ -1,13 +1,18 @@
-import React, { useState, useEffect, useRef, useCallback } from 'react';
-import { Heart, Calendar, MessageCircle, Clock, Sparkles, Camera, MapPin, Play, Coins, Plus, X, User, Pause, StickyNote, ChevronDown, ChevronUp, Send, Check, MessageSquareHeart, Trash2, Pencil } from 'lucide-react';
+import { useState, useEffect, useCallback } from 'react';
+import { Calendar, MessageCircle, Clock, MapPin, Play, Coins, User, StickyNote, MessageSquareHeart, Trash2, Pencil } from 'lucide-react';
 import SettingsView from './components/SettingsView';
 import UpgradeView, { BillingResultView } from './components/UpgradeView';
 import RoleplayView from './components/RoleplayView';
 import ScriptUploadModal from './components/ScriptUploadModal';
+import GamesView from './components/GamesView';
+import CoinShopView from './components/CoinShopView';
+import OurJourneyView from './components/OurJourneyView';
+import AuthModal from './components/AuthModal';
+import CalendarView from './components/CalendarView';
+import ConflictView from './components/ConflictView';
 import WallView from './components/WallView';
 import EventsView from './components/EventsView';
 import type { WallExample } from './components/WallPostComposer';
-import { AchievementsView, IntimacyStatsCards, CalendarHeatmap } from './components/AchievementsView';
 import Header from './components/Header';
 import { NotificationContainer } from './components/ErrorNotification';
 import IntimacyRequestsHistory from './components/IntimacyRequestsHistory';
@@ -16,14 +21,12 @@ import NotificationInbox from './components/NotificationInbox';
 import PairingInvitationHandler from './components/PairingInvitationHandler';
 import { apiService, getTokenExpiry, clearAuthStorage } from './services/api';
 import type { CycleRecord } from './services/api';
-import { periodDateSet, fertileDateSet, predictedPeriodDateSet } from './utils/cycle';
-import { getPrimaryTimezone, formatMonthYear, formatYmdInTz, browserTz } from './utils/datetime';
+import { getPrimaryTimezone, formatYmdInTz, browserTz } from './utils/datetime';
 import { TimezoneProvider } from './contexts/TimezoneContext';
-import { conflictPhraseTiers } from './data/conflictSteps';
 import { useScrollLock } from './hooks/useScrollLock';
 import { usePageTracking } from './hooks/usePageTracking';
 
-interface IntimateRecord {
+export interface IntimateRecord {
   id: number;
   apiId?: string;
   date: string;
@@ -45,7 +48,7 @@ interface Nicknames {
   partner2: string;
 }
 
-interface JourneyMilestone {
+export interface JourneyMilestone {
   id: string;
   type: 'meeting' | 'first_date' | 'first_kiss' | 'first_sex' | 'marriage' | 'child_born' | 'intimacy_milestone' | 'custom';
   date: string;
@@ -57,7 +60,7 @@ interface JourneyMilestone {
   isCustom?: boolean;
 }
 
-interface Notification {
+export interface Notification {
   id: string;
   type: 'success' | 'error' | 'info' | 'warning';
   title: string;
@@ -115,7 +118,7 @@ interface ApiCustomGift {
   createdBy?: string;
 }
 
-interface CoinGift {
+export interface CoinGift {
   id: string;
   title: string;
   description: string;
@@ -143,7 +146,7 @@ interface User {
   createdAt: string;
 }
 
-interface AuthState {
+export interface AuthState {
   user: User | null;
   isAuthenticated: boolean;
   partnerConnected: boolean;
@@ -384,7 +387,7 @@ const defaultWallExamples: WallExample[] = [
 ];
 
 // Add interfaces for the data structures
-interface ForeplayActivity {
+export interface ForeplayActivity {
   title: string;
   description: string;
   duration: string;
@@ -392,7 +395,7 @@ interface ForeplayActivity {
   tips: string[];
 }
 
-interface PositionSuggestion {
+export interface PositionSuggestion {
   name: string;
   difficulty: string;
   description: string;
@@ -618,44 +621,6 @@ const LoveTimeApp = () => {
     const timer = setTimeout(() => setShowTagline(false), 10000);
     return () => clearTimeout(timer);
   }, []);
-
-  // File input refs
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  // Image compression utility
-  const compressImage = (file: File, maxWidth: number = 800, maxHeight: number = 600, quality: number = 0.8): Promise<string> => {
-    return new Promise((resolve) => {
-      const canvas = document.createElement('canvas');
-      const ctx = canvas.getContext('2d');
-      const img = new Image();
-      
-      img.onload = () => {
-        // Calculate new dimensions
-        let { width, height } = img;
-        if (width > height) {
-          if (width > maxWidth) {
-            height = (height * maxWidth) / width;
-            width = maxWidth;
-          }
-        } else {
-          if (height > maxHeight) {
-            width = (width * maxHeight) / height;
-            height = maxHeight;
-          }
-        }
-        
-        canvas.width = width;
-        canvas.height = height;
-        
-        // Draw and compress
-        ctx?.drawImage(img, 0, 0, width, height);
-        const compressedDataUrl = canvas.toDataURL('image/jpeg', quality);
-        resolve(compressedDataUrl);
-      };
-      
-      img.src = URL.createObjectURL(file);
-    });
-  };
 
   // Authentication functions
   const generatePartnerCode = () => {
@@ -1769,315 +1734,6 @@ const LoveTimeApp = () => {
 
 
 
-  const romanticGames = [
-    { 
-      title: '真心話大冒險', 
-      desc: '輪流問對方從未問過的問題，增進了解',
-      instructions: [
-        '1. 準備紙條寫下問題，放入盒子中',
-        '2. 輪流抽取問題，必須誠實回答',
-        '3. 問題可以包括：夢想、恐懼、幻想、回憶',
-        '4. 大冒險可以是：唱歌、按摩、親吻、擥抱',
-        '5. 創造安全舒適的環境，互相尊重'
-      ],
-      questions: [
-        '如果今晚是世界末日，你最想和我做什麼？',
-        '你對我身體最著迷的部位是哪裡？',
-        '描述你心中最完美的約會夜晚',
-        '你最想在哪個地方和我親熱？',
-        '如果可以實現一個性幻想，會是什麼？'
-      ]
-    },
-    { 
-      title: '感官按摩', 
-      desc: '用觸覺喚醒彼此的感官，放鬆身心',
-      instructions: [
-        '1. 準備香薰蠟燭和按摩油',
-        '2. 調暗燈光，播放輕柔音樂',
-        '3. 從肩膀開始，慢慢按摩全身',
-        '4. 專注於對方的反應和呼吸',
-        '5. 輪流為對方按摩，享受被愛撫的感覺'
-      ],
-      tips: [
-        '用溫熱的按摩油，避免太冷',
-        '變化按摩的力度和節奏',
-        '不要忽略敏感部位',
-        '用指尖輕撫，製造酥麻感',
-        '按摩時保持眼神交流'
-      ]
-    },
-    { 
-      title: '情慾骰子', 
-      desc: '用骰子決定親密動作，增加刺激感',
-      instructions: [
-        '1. 準備兩個骰子（動作骰子和部位骰子）',
-        '2. 動作包括：親吻、撫摸、按摩、舔舐、輕咬、吹氣',
-        '3. 部位包括：唇、頸、耳、胸、腰、腿',
-        '4. 輪流擲骰子，按照結果執行',
-        '5. 可以設定時間限制，增加刺激感'
-      ],
-      variations: [
-        '加入溫度元素：冰塊或溫水',
-        '使用羽毛或絲巾增加質感',
-        '設定不同的強度等級',
-        '加入調情話語的要求',
-        '結合不同的姿勢或位置'
-      ]
-    },
-    { 
-      title: '慢燃調情', 
-      desc: '用語言和眼神慢慢點燃激情',
-      instructions: [
-        '1. 面對面坐著，保持眼神接觸',
-        '2. 輪流說出對方最性感的特質',
-        '3. 描述你想對對方做的事情',
-        '4. 用手輕撫但不要碰觸私密部位',
-        '5. 建立期待感，延遲滿足'
-      ],
-      phrases: [
-        '你知道嗎？當你看著我的時候...',
-        '我一直在想像...',
-        '今晚我想要...',
-        '你讓我感到...',
-        '如果我們現在...'
-      ]
-    },
-    { 
-      title: '記憶重現', 
-      desc: '重演你們最難忘的親密時刻',
-      instructions: [
-        '1. 分享彼此最難忘的親密回憶',
-        '2. 選擇一個想要重演的場景',
-        '3. 儘可能還原當時的情境',
-        '4. 加入新的元素讓體驗更豐富',
-        '5. 專注於當時的感覺和情緒'
-      ],
-      scenarios: [
-        '第一次親吻的地點和感覺',
-        '最浪漫的一次約會夜晚',
-        '最激情的一次親密時光',
-        '最意外的親密時刻',
-        '最溫柔纏綿的早晨'
-      ]
-    },
-    {
-      title: '慾望清單',
-      desc: '分享彼此的性幻想和願望',
-      instructions: [
-        '1. 各自寫下5個親密願望',
-        '2. 交換清單，討論每一項',
-        '3. 選擇雙方都感興趣的項目',
-        '4. 制定實現這些願望的計劃',
-        '5. 從最容易實現的開始嘗試'
-      ],
-      categories: [
-        '新的地點或環境',
-        '不同的角色扮演',
-        '新的親密方式',
-        '感官刺激體驗',
-        '浪漫情境設定'
-      ]
-    },
-    {
-      title: '回憶倒帶',
-      desc: 'App 隨機抽一題溫柔回憶，說完後讓另一方做一個「回饋動作」',
-      instructions: [
-        '1. 一人從下方問題中隨機抽一題',
-        '2. 認真回想，誠實分享你的答案',
-        '3. 另一方靜靜聽完，選擇一個回饋動作回應',
-        '4. 回饋動作：摸頭、擁抱、輕靠肩，或任何溫柔的肢體互動',
-        '5. 輪流換邊抽題，慢慢回顧你們的點滴'
-      ],
-      questions: [
-        '第一次覺得對方很可愛是什麼時候？',
-        '最近一次被對方感動是因為？',
-        '對方做過最讓你心動的小事是什麼？',
-        '想到我們最幸福的一天，你會想到哪一天？',
-        '對方哪一個習慣讓你忍不住微笑？',
-        ...customMemoryQuestions
-      ]
-    },
-    {
-      title: '情緒模仿秀',
-      desc: '抽到一個情緒，不能說話，只能用身體或眼神表演，對方猜',
-      instructions: [
-        '1. 一人從下方情緒中隨機抽一個',
-        '2. 不能說話，只能用身體動作、表情、眼神演出來',
-        '3. 對方在 60 秒內猜出情緒',
-        '4. 猜對換邊；猜錯也好玩，互相回饋哪裡演得像、哪裡可以再加強',
-        '5. 過程輕鬆，氣氛帶點曖昧'
-      ],
-      emotions: [
-        '害羞',
-        '惹人疼',
-        '想撒嬌',
-        '吃醋',
-        '欲拒還迎',
-        '故作冷淡',
-        ...customEmotions
-      ]
-    }
-  ];
-
-  const conflictResolutions = [
-    { title: '傾聽練習', desc: '給對方5分鐘不被打斷的表達時間' },
-    { title: '愛的語言', desc: '用"我感覺"而不是"你總是"來表達' },
-    { title: '擁抱和解', desc: '爭吵後先給對方一個溫暖的擁抱' },
-    { title: '寫信溝通', desc: '將想說的話寫成信，避免激烈爭吵' },
-    { title: '約定時間', desc: '設定專門的溝通時間，心平氣和討論' }
-  ];
-
-  // Foreplay Activities
-  const foreplayActivities = [
-    {
-      title: '感官按摩',
-      description: '用溫熱的按摩油為彼此按摩，喚醒身體的每一寸肌膚',
-      duration: '15-30分鐘',
-      coins: 150,
-      tips: ['使用香薰蠟燭營造氛圍', '從肩膀開始慢慢按摩', '專注於對方的反應']
-    },
-    {
-      title: '冰火兩重天',
-      description: '用冰塊和溫水交替刺激敏感部位，帶來不同的感官體驗',
-      duration: '10-20分鐘',
-      coins: 200,
-      tips: ['準備冰塊和溫水', '輕柔地在身體上滑動', '注意對方的感受']
-    },
-    {
-      title: '羽毛撫摸',
-      description: '用柔軟的羽毛輕撫身體，帶來酥麻的快感',
-      duration: '10-15分鐘',
-      coins: 150,
-      tips: ['選擇柔軟的羽毛', '從不敏感部位開始', '變化力度和速度']
-    },
-    {
-      title: '蒙眼遊戲',
-      description: '蒙上眼睛，讓其他感官更加敏銳，增加神秘感',
-      duration: '20-30分鐘',
-      coins: 250,
-      tips: ['使用絲巾或眼罩', '專注於觸覺和聽覺', '保持溝通']
-    }
-  ];
-
-  // Position suggestions
-  const positionSuggestions = [
-    {
-      name: '蓮花式',
-      difficulty: '簡單',
-      description: '面對面坐著，增進親密感和眼神交流',
-      coins: 200,
-      benefits: ['增進情感連結', '便於親吻和撫摸', '適合慢節奏']
-    },
-    {
-      name: '側臥式',
-      difficulty: '簡單',
-      description: '側躺進行，適合長時間親密',
-      coins: 200,
-      benefits: ['減少疲勞', '適合懷孕期', '便於撫摸']
-    },
-    {
-      name: '後入式',
-      difficulty: '中等',
-      description: '從後方進入，帶來不同的刺激感',
-      coins: 300,
-      benefits: ['深度刺激', '便於撫摸敏感部位', '角度變化']
-    },
-    {
-      name: '站立式',
-      difficulty: '困難',
-      description: '站立進行，增加新鮮感和刺激',
-      coins: 400,
-      benefits: ['新鮮體驗', '不同角度', '增加難度挑戰']
-    },
-    {
-      name: '開瓶器式',
-      difficulty: '中等',
-      description: '側躺在床或長凳的邊緣，大腿併攏，伴侶從你身後站立進入。',
-      coins: 300,
-      benefits: ['雙腿併攏使緊緻感更強', '便於伴侶從身後深入', '進階：用屁股配合伴侶的節奏']
-    },
-    {
-      name: '捲餅沾醬式',
-      difficulty: '中等',
-      description: '右側躺下，伴侶跨坐在你的右腿上，左腿環繞在他身體的左側。',
-      coins: 300,
-      benefits: ['深度插入的同時保持目光接觸', '結合狗狗式的深度與面對面的親密', '進階：讓伴侶摩擦你的陰蒂']
-    },
-    {
-      name: '女牛仔騎乘位',
-      difficulty: '中等',
-      description: '跪在伴侶身上，按壓他的胸部，沿著他的大腿上下滑動；他可以抓住你的大腿或臀部抬起插入。',
-      coins: 300,
-      benefits: ['伴侶分擔體重，緩解雙腿疲勞', '女性主導，更容易達到高潮', '能延遲男性高潮，雙贏體驗']
-    },
-    {
-      name: '獨輪車式',
-      difficulty: '困難',
-      description: '手腳並用撐地，伴侶從骨盆向下抱住你，腰部放在你的大腿之間。',
-      coins: 400,
-      benefits: ['插入更深，刺激更強', '同時鍛鍊手臂力量', '挑戰耐力與默契']
-    },
-    {
-      name: '背面騎乘位',
-      difficulty: '中等',
-      description: '伴侶仰臥，你反向跨坐在他身上，朝向他的腳。',
-      coins: 300,
-      benefits: ['女性主導節奏與深度', '可以引導伴侶學習你喜歡的節奏', '不同角度的視覺刺激']
-    },
-    {
-      name: '拱門式',
-      difficulty: '簡單',
-      description: '面對面坐下，雙腿伸直，膝蓋放在對方大腿上，上半身都稍微向後傾斜。',
-      coins: 200,
-      benefits: ['能看到對方整個身體', '自由控制插入的深度、速度與角度', '進階：身體越後傾，G 點刺激越強']
-    },
-    {
-      name: '彈珠大師式',
-      difficulty: '困難',
-      description: '用肩膀支撐身體做半橋式，伴侶以跪姿插入。',
-      coins: 400,
-      benefits: ['對伴侶來說姿勢舒適', '能同時刺激陰蒂並按摩陰阜', '挑戰核心與肩膀力量']
-    }
-  ];
-
-  // 連續組合技：一次連續換多個姿勢的進階挑戰。完成可獲得額外金幣獎金（高於單獨嘗試各姿勢的總和）。
-  const comboSuggestions = [
-    {
-      name: '溫柔三部曲',
-      difficulty: '簡單',
-      description: '從親密接觸慢慢加深連結，適合放鬆的夜晚。',
-      sequence: ['蓮花式', '側臥式', '後入式'],
-      bonusCoins: 900,
-    },
-    {
-      name: '漸進挑戰',
-      difficulty: '中等',
-      description: '由淺入深、節奏漸強的中階組合，考驗默契。',
-      sequence: ['側臥式', '後入式', '站立式'],
-      bonusCoins: 1200,
-    },
-    {
-      name: '大膽探索',
-      difficulty: '困難',
-      description: '高強度組合，挑戰你們的耐力與創意。',
-      sequence: ['站立式', '後入式', '蓮花式'],
-      bonusCoins: 1500,
-    },
-  ];
-
-
-  // Chinese-numeral date formatter for editorial meta lines. e.g. 二〇二五年九月七日
-  const toChineseNum = (n: number): string => {
-    const cn = ['〇', '一', '二', '三', '四', '五', '六', '七', '八', '九'];
-    if (n < 10) return cn[n];
-    if (n < 20) return n === 10 ? '十' : `十${cn[n - 10]}`;
-    if (n < 30) return n === 20 ? '二十' : `二十${cn[n - 20]}`;
-    return n === 30 ? '三十' : `三十${cn[n - 30]}`;
-  };
-  const chineseYear = (y: number): string =>
-    y.toString().split('').map(c => '〇一二三四五六七八九'[parseInt(c, 10)]).join('');
-  const formatChineseDate = (d: Date): string =>
-    `${chineseYear(d.getFullYear())}年${toChineseNum(d.getMonth() + 1)}月${toChineseNum(d.getDate())}日`;
 
   // "Together since" date — prefers the user-set 交往日期 milestone, then
   // earliest intimate record, then the account creation date.
@@ -2098,3132 +1754,14 @@ const LoveTimeApp = () => {
     ? Math.max(1, Math.floor((Date.now() - togetherSince.getTime()) / 86400000) + 1)
     : 0;
 
-  const CalendarView = () => {
-    const getCurrentTime = () => {
-      const now = new Date();
-      const hours = now.getHours().toString().padStart(2, '0');
-      const minutes = now.getMinutes().toString().padStart(2, '0');
-      return `${hours}:${minutes}`;
-    };
-
-    const [recordForm, setRecordForm] = useState(() => {
-      if (editingRecord) {
-        return {
-          date: editingRecord.date,
-          time: editingRecord.time,
-          mood: editingRecord.mood,
-          notes: editingRecord.notes || '',
-          description: editingRecord.description || '',
-          duration: editingRecord.duration || '',
-          location: editingRecord.location || '',
-          photo: editingRecord.photo || '',
-          roleplayScript: editingRecord.roleplayScript || ''
-        };
-      }
-      return {
-        date: selectedDate,
-        time: getCurrentTime(),
-        mood: '💕',
-        notes: '',
-        description: '',
-        duration: '',
-        location: '',
-        photo: '',
-        roleplayScript: ''
-      };
-    });
-
-    useEffect(() => {
-      if (editingRecord && showRecordModal) {
-        setRecordForm({
-          date: editingRecord.date,
-          time: editingRecord.time,
-          mood: editingRecord.mood,
-          notes: editingRecord.notes || '',
-          description: editingRecord.description || '',
-          duration: editingRecord.duration || '',
-          location: editingRecord.location || '',
-          photo: editingRecord.photo || '',
-          roleplayScript: editingRecord.roleplayScript || ''
-        });
-      }
-    }, [editingRecord, showRecordModal]);
-
-    const recordsByDate = React.useMemo(() => {
-      const map = new Map<string, IntimateRecord[]>();
-      intimateRecords.forEach(r => {
-        const existing = map.get(r.date) || [];
-        existing.push(r);
-        map.set(r.date, existing);
-      });
-      return map;
-    }, [intimateRecords]);
-
-    const [recordType, setRecordType] = useState<'intimacy' | 'period'>('intimacy');
-    const [periodLengthDays, setPeriodLengthDays] = useState<number>(5);
-
-    const cycleEnabled = !!authState.user?.cycle_tracking_enabled;
-    const periodDates = React.useMemo(() => cycleEnabled ? periodDateSet(cycleRecords) : undefined, [cycleEnabled, cycleRecords]);
-    const fertileDates = React.useMemo(() => cycleEnabled ? fertileDateSet(cycleRecords) : undefined, [cycleEnabled, cycleRecords]);
-    const predictedPeriodDates = React.useMemo(() => cycleEnabled ? predictedPeriodDateSet(cycleRecords) : undefined, [cycleEnabled, cycleRecords]);
-
-    const handlePhotoUpload = async (event: React.ChangeEvent<HTMLInputElement>) => {
-      const file = event.target.files?.[0];
-      if (!file) return;
-
-      try {
-        const compressedImage = await compressImage(file);
-        setRecordForm({...recordForm, photo: compressedImage});
-      } catch (err) {
-        console.error('Failed to upload photo:', err);
-        showNotification({
-          type: 'error',
-          title: '照片上傳失敗',
-          message: '請稍後再試',
-          duration: 3000
-        });
-      }
-    };
-
-    const handleSubmitRecord = async () => {
-      if (!recordForm.date) {
-        showNotification({ type: 'error', title: '驗證錯誤', message: '請選擇日期', duration: 6000 });
-        return;
-      }
-
-      if (recordType === 'period') {
-        try {
-          const created = await apiService.createCycleRecord({
-            startDate: recordForm.date,
-            lengthDays: periodLengthDays,
-            notes: recordForm.notes || undefined,
-          });
-          setCycleRecords(prev => [created, ...prev]);
-          showNotification({ type: 'success', title: '已儲存', message: '週期紀錄已新增', duration: 3000 });
-        } catch (error: unknown) {
-          showNotification({ type: 'error', title: '建立失敗', message: (error as Error)?.message || '無法建立週期紀錄', duration: 5000 });
-          return;
-        }
-        setShowRecordModal(false);
-        setEditingRecord(null);
-        setRecordType('intimacy');
-        return;
-      }
-
-      if (!recordForm.time) {
-        showNotification({ type: 'error', title: '驗證錯誤', message: '請選擇時間', duration: 6000 });
-        return;
-      }
-
-      if (editingRecord) {
-        try {
-          const updates: Partial<IntimateRecord> = {
-            date: recordForm.date,
-            time: recordForm.time,
-            mood: recordForm.mood,
-            notes: recordForm.notes || undefined,
-            description: recordForm.description || undefined,
-            duration: recordForm.duration || undefined,
-            location: recordForm.location || undefined,
-            roleplayScript: recordForm.roleplayScript || undefined,
-          };
-          await apiService.updateIntimateRecord(editingRecord.apiId!, updates);
-          setIntimateRecords(prev => prev.map(r =>
-            r.id === editingRecord.id ? { ...r, ...updates } : r
-          ));
-          showNotification({ type: 'success', title: '已更新', message: '記錄已成功更新', duration: 3000 });
-        } catch (error: unknown) {
-          console.error('Error updating record:', error);
-          showNotification({ type: 'error', title: '更新失敗', message: (error as Error)?.message || '無法更新記錄', duration: 5000 });
-          return;
-        }
-      } else {
-        await addIntimateRecord(
-          recordForm.date,
-          recordForm.time,
-          recordForm.mood,
-          recordForm.notes || undefined,
-          recordForm.photo,
-          recordForm.description || undefined,
-          recordForm.duration || undefined,
-          recordForm.location || undefined,
-          recordForm.roleplayScript || undefined
-        );
-      }
-
-      setShowRecordModal(false);
-      setEditingRecord(null);
-      setRecordForm({
-        date: selectedDate,
-        time: getCurrentTime(),
-        mood: '💕',
-        notes: '',
-        description: '',
-        duration: '',
-        location: '',
-        photo: '',
-        roleplayScript: ''
-      });
-    };
-
-    return (
-      <div className="space-y-10">
-        <div className="border-b border-petal-rule pb-7">
-          <div className="flex flex-col md:flex-row md:items-end md:justify-between gap-6">
-            <div>
-              <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-                — A · 記錄
-              </div>
-              <h2 className="font-display text-4xl md:text-5xl font-light tracking-tight text-petal-ink leading-[1.05]">
-                記錄<em className="not-italic font-light italic text-pink-600">時光</em>
-              </h2>
-            </div>
-            <div className="text-left md:text-right font-body text-sm text-petal-muted leading-relaxed">
-              {togetherSince ? (
-                <>
-                  <strong className="block text-petal-ink font-display font-semibold text-base tracking-tight mb-0.5">
-                    {daysTogether} days together
-                  </strong>
-                  自{formatChineseDate(togetherSince)}
-                </>
-              ) : (
-                <span className="font-display italic">— 開始你們的旅程 —</span>
-              )}
-            </div>
-          </div>
-          <div className="mt-6 flex flex-col md:flex-row gap-3 md:items-center">
-            <div className="flex items-center gap-3 md:flex-1">
-              <span className="font-body text-xs uppercase tracking-[0.14em] text-petal-muted">日期</span>
-              <input
-                type="date"
-                value={selectedDate}
-                onChange={(e) => setSelectedDate(e.target.value)}
-                className="flex-1 max-w-xs px-4 py-2.5 border border-petal-rule rounded-md bg-white font-body text-sm text-petal-ink focus:outline-none focus:border-petal-rose-deep transition-colors"
-              />
-            </div>
-            <button
-              data-testid="add-record-button"
-              onClick={() => {
-                setEditingRecord(null);
-                setRecordForm({date: selectedDate, time: getCurrentTime(), mood: '💕', notes: '', description: '', duration: '', location: '', photo: '', roleplayScript: ''});
-                setShowRecordModal(true);
-              }}
-              className="px-6 py-2.5 bg-petal-ink text-petal-cream rounded-md font-display italic text-base hover:bg-pink-700 transition-colors"
-            >
-              + 為這一天添一筆
-            </button>
-          </div>
-        </div>
-
-        {/* Intimacy Stats — 4 cards */}
-        <IntimacyStatsCards
-          records={intimateRecords}
-          birthDate={authState.user?.birth_date}
-          onOpenSettings={() => setCurrentView('settings')}
-        />
-
-        {/* Enhanced Record Modal */}
-        {showRecordModal && (
-          <div className="fixed inset-0 bg-petal-ink/40 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-            <div className="bg-petal-cream rounded-md shadow-petal max-w-2xl w-full max-h-[min(90vh,calc(100dvh-80px))] overflow-y-auto overscroll-contain border border-petal-rule">
-              <div className="p-5 sm:p-6">
-                <div className="flex justify-between items-end mb-5 pb-4 border-b border-petal-rule">
-                  <div>
-                    <div className="font-body text-[11px] font-medium uppercase tracking-[0.16em] text-petal-muted mb-2">
-                      — {editingRecord ? '編輯記錄' : '新的記錄'}
-                    </div>
-                    <h3 data-testid="record-modal-heading" className="font-display text-3xl font-light tracking-tight text-petal-ink">
-                      {editingRecord ? '編輯' : '記錄'}<em className="not-italic font-light italic text-pink-600">{recordType === 'period' ? '月經' : '親密時光'}</em>
-                    </h3>
-                  </div>
-                  <button
-                    onClick={() => { setShowRecordModal(false); setEditingRecord(null); }}
-                    data-testid="record-modal-close-button"
-                    className="text-petal-muted hover:text-petal-ink text-2xl font-light transition-colors leading-none"
-                  >
-                    ×
-                  </button>
-                </div>
-
-                <div className="space-y-3">
-                  {/* Event type — only when cycle tracking is opted in and creating a new record */}
-                  {cycleEnabled && !editingRecord && (
-                    <div>
-                      <label className="block text-sm font-medium text-gray-700 mb-2">事件類型</label>
-                      <div className="inline-flex rounded-md border border-petal-rule overflow-hidden" role="tablist">
-                        <button
-                          type="button"
-                          data-testid="record-type-intimacy"
-                          onClick={() => setRecordType('intimacy')}
-                          className={`px-4 py-2 text-sm font-display italic transition-colors ${recordType === 'intimacy' ? 'bg-petal-ink text-petal-cream' : 'bg-white text-petal-ink hover:bg-petal-cream-2'}`}
-                        >親密時光</button>
-                        <button
-                          type="button"
-                          data-testid="record-type-period"
-                          onClick={() => setRecordType('period')}
-                          className={`px-4 py-2 text-sm font-display italic transition-colors border-l border-petal-rule ${recordType === 'period' ? 'bg-red-500 text-white' : 'bg-white text-petal-ink hover:bg-petal-cream-2'}`}
-                        >月經</button>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Basic Info */}
-                  <div className="space-y-3">
-                    <div>
-                      <label className="block text-sm font-medium text-gray-700 mb-2">{recordType === 'period' ? '週期開始日' : '日期選擇'}</label>
-                      <CalendarDatePicker
-                        selectedDate={recordForm.date}
-                        onDateSelect={(date) => setRecordForm({...recordForm, date})}
-                      />
-                    </div>
-                    {recordType === 'period' && (
-                      <div>
-                        <label className="block text-sm font-medium text-gray-700 mb-2">月經天數</label>
-                        <input
-                          type="number"
-                          min={1}
-                          max={14}
-                          value={periodLengthDays}
-                          onChange={(e) => {
-                            const v = parseInt(e.target.value, 10);
-                            if (!Number.isNaN(v)) setPeriodLengthDays(Math.min(14, Math.max(1, v)));
-                          }}
-                          data-testid="record-period-length-input"
-                          className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                        />
-                        <p className="text-xs text-petal-muted mt-1">1–14 天，預設 5 天。</p>
-                      </div>
-                    )}
-                  </div>
-
-                  {recordType !== 'period' && (
-                  <>
-                  {/* Photo Upload — compact */}
-                  <div>
-                    <div className="flex items-center gap-3 flex-wrap">
-                      <button
-                        type="button"
-                        onClick={() => fileInputRef.current?.click()}
-                        data-testid="record-photo-upload-button"
-                        className="inline-flex items-center gap-2 px-3 py-2 border border-petal-rule rounded-md text-sm text-petal-ink hover:bg-petal-cream-2 transition-colors"
-                      >
-                        <Camera className="w-4 h-4" />
-                        {recordForm.photo ? '更換照片' : '上傳照片 (可選)'}
-                      </button>
-                      <input
-                        ref={fileInputRef}
-                        type="file"
-                        accept="image/*"
-                        onChange={handlePhotoUpload}
-                        className="hidden"
-                      />
-                      {recordForm.photo && (
-                        <div className="relative">
-                          <img
-                            src={recordForm.photo}
-                            alt="記憶照片"
-                            className="w-16 h-16 object-cover rounded-md border border-petal-rule"
-                          />
-                          <button
-                            type="button"
-                            onClick={() => setRecordForm({...recordForm, photo: ''})}
-                            aria-label="移除照片"
-                            className="absolute -top-1.5 -right-1.5 bg-red-500 text-white rounded-full w-5 h-5 text-xs leading-none flex items-center justify-center"
-                          >
-                            ×
-                          </button>
-                        </div>
-                      )}
-                    </div>
-                  </div>
-
-                  {/* Description and Details */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">描述你們做了什麼</label>
-                    <textarea
-                      value={recordForm.description}
-                      onChange={(e) => setRecordForm({...recordForm, description: e.target.value})}
-                      placeholder="分享這個美好時光的細節..."
-                      data-testid="record-description-input"
-                      className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500 h-16"
-                    />
-                  </div>
-
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">
-                      <MapPin className="w-4 h-4 inline mr-2" />
-                      地點
-                    </label>
-                    <input
-                      type="text"
-                      value={recordForm.location}
-                      onChange={(e) => setRecordForm({...recordForm, location: e.target.value})}
-                      placeholder="例如：臥室、客廳"
-                      data-testid="record-location-input"
-                      className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                    />
-                  </div>
-
-                  {/* Roleplay Script Reference */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">
-                      <Play className="w-4 h-4 inline mr-2" />
-                      角色扮演劇本 (可選)
-                    </label>
-                    <select
-                      value={recordForm.roleplayScript}
-                      onChange={(e) => setRecordForm({...recordForm, roleplayScript: e.target.value})}
-                      data-testid="record-roleplay-select"
-                      className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                    >
-                      <option value="">未使用劇本</option>
-                      {/* Default scripts */}
-                      {defaultRoleplayScripts.map((script, index) => (
-                        <option key={`default-${index}`} value={script.title}>{script.title}</option>
-                      ))}
-                      {/* Custom scripts */}
-                      {customScripts.map((script, index) => (
-                        <option key={`custom-${index}`} value={script.title}>{script.title} (自定義)</option>
-                      ))}
-                    </select>
-                  </div>
-
-                  </>
-                  )}
-
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">備註</label>
-                    <textarea
-                      value={recordForm.notes}
-                      onChange={(e) => setRecordForm({...recordForm, notes: e.target.value})}
-                      placeholder="記錄這個特別時刻的感受..."
-                      data-testid="record-notes-input"
-                      className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500 h-16"
-                    />
-                  </div>
-                </div>
-
-                <div className="flex space-x-3 mt-6 pt-4 border-t border-petal-rule">
-                  <button
-                    onClick={() => { setShowRecordModal(false); setEditingRecord(null); }}
-                    data-testid="record-cancel-button"
-                    className="flex-1 px-4 py-3 border border-petal-rule text-petal-ink rounded-md hover:bg-petal-cream-2 transition-colors font-body text-sm"
-                  >
-                    取消
-                  </button>
-                  <button
-                    onClick={handleSubmitRecord}
-                    data-testid="record-submit-button"
-                    className="flex-1 px-4 py-3 bg-petal-ink text-petal-cream rounded-md hover:bg-pink-700 transition-colors font-display italic text-base"
-                  >
-                    {editingRecord ? '更新記錄' : '保存記錄'}
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-
-        {/* Monthly Calendar Distribution */}
-        <div>
-          <div className="flex items-baseline justify-between mb-6">
-            <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink">
-              月度<em className="not-italic font-light italic text-pink-600">記錄分佈</em>
-            </h3>
-          </div>
-          <div className="bg-white rounded-md border border-petal-rule p-5 sm:p-6">
-            <CalendarHeatmap
-              data={intimateRecords}
-              year={calendarMonth.getFullYear()}
-              month={calendarMonth.getMonth()}
-              title=""
-              showMonthLabels={true}
-              periodDates={periodDates}
-              predictedPeriodDates={predictedPeriodDates}
-              fertileDates={fertileDates}
-              onNavigate={(y, m) => setCalendarMonth(new Date(y, m, 1))}
-              onDaySelect={(day) => {
-                const dayRecords = recordsByDate.get(day) || [];
-                if (dayRecords.length === 0) {
-                  setEditingRecord(null);
-                  setSelectedDate(day);
-                  setShowRecordModal(true);
-                } else if (dayRecords.length === 1) {
-                  setSelectedRecord(dayRecords[0]);
-                  setShowRecordDetail(true);
-                } else {
-                  setDayPickerDate(day);
-                  setDayPickerRecords(
-                    dayRecords.slice().sort((a, b) =>
-                      (b.date + 'T' + b.time).localeCompare(a.date + 'T' + a.time)
-                    )
-                  );
-                }
-              }}
-            />
-          </div>
-        </div>
-
-        {/* Record List */}
-        <div>
-          <div className="flex items-baseline justify-between mb-6">
-            <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink">
-              親密<em className="not-italic font-light italic text-pink-600">記錄</em>
-            </h3>
-            <span className="font-display italic font-light text-sm text-petal-muted">
-              共 <b className="not-italic font-normal text-petal-ink">{intimateRecords.length}</b> 次
-            </span>
-          </div>
-          <div className="max-h-[28rem] overflow-y-auto overflow-x-hidden">
-            {(() => {
-              const filtered = intimateRecords
-                .slice()
-                .sort((a, b) => (b.date + 'T' + b.time).localeCompare(a.date + 'T' + a.time));
-              return filtered.length > 0 ? filtered.map((record, idx) => (
-                <article
-                  key={record.id}
-                  onClick={() => showRecordDetails(record.id)}
-                  className={`group grid grid-cols-[40px_1fr_auto] gap-5 py-5 cursor-pointer hover:bg-petal-cream-2/40 -mx-2 px-2 transition-colors ${
-                    idx === 0 ? '' : 'border-t border-petal-rule-soft'
-                  }`}
-                >
-                  <div className="text-base opacity-70 saturate-75 mt-1 text-center leading-none">
-                    {record.mood}
-                  </div>
-                  <div className="min-w-0">
-                    <div className="font-display italic font-light text-sm text-petal-muted mb-1">
-                      {record.date} · {record.time}
-                    </div>
-                    {record.description && (
-                      <p className="font-body text-[15px] leading-relaxed text-petal-ink mb-1.5">
-                        {record.description}
-                      </p>
-                    )}
-                    <div className="flex flex-wrap gap-2 text-[11px] text-petal-muted mt-1.5">
-                      {record.duration && (
-                        <span className="inline-flex items-start max-w-full px-2.5 py-0.5 border border-petal-rule rounded-md">
-                          <Clock className="w-3 h-3 mr-1 mt-[3px] flex-shrink-0" />
-                          <span className="break-words leading-snug">{record.duration}</span>
-                        </span>
-                      )}
-                      {record.location && (
-                        <span className="inline-flex items-start max-w-full px-2.5 py-0.5 border border-petal-rule rounded-md">
-                          <MapPin className="w-3 h-3 mr-1 mt-[3px] flex-shrink-0" />
-                          <span className="break-words leading-snug">{record.location}</span>
-                        </span>
-                      )}
-                      {record.roleplayScript && (
-                        <span className="inline-flex items-start max-w-full px-2.5 py-0.5 border border-petal-sage/60 bg-petal-sage/10 text-petal-sage-deep rounded-md">
-                          <Play className="w-3 h-3 mr-1 mt-[3px] flex-shrink-0" />
-                          <span className="break-words leading-snug">{record.roleplayScript}</span>
-                        </span>
-                      )}
-                    </div>
-                    {record.notes && (
-                      <p className="font-display italic font-light text-sm text-petal-ink-soft mt-2.5 pl-3 border-l border-petal-rose-soft leading-relaxed">
-                        "{record.notes}"
-                      </p>
-                    )}
-                    {record.photo && (
-                      <img
-                        src={record.photo}
-                        alt="記憶照片"
-                        className="mt-3 w-24 h-24 rounded-md object-cover border border-petal-rule"
-                      />
-                    )}
-                  </div>
-                  <div className="flex items-start pt-1">
-                    <button
-                      onClick={(e) => { e.stopPropagation(); openDeleteConfirm(record); }}
-                      className="opacity-0 group-hover:opacity-100 text-petal-muted hover:text-red-500 transition-all p-1 rounded"
-                    >
-                      <Trash2 className="w-4 h-4" />
-                    </button>
-                  </div>
-                </article>
-              )) : (
-                <div className="border border-dashed border-petal-rule rounded-md py-10 px-6 text-center">
-                  <p className="font-display italic font-light text-base text-petal-muted">
-                    還沒有記錄 — 開始你們的愛情之旅吧
-                  </p>
-                </div>
-              );
-            })()}
-          </div>
-        </div>
-
-        <div className="border-t border-petal-rule pt-10">
-          <AchievementsView />
-        </div>
-      </div>
-    );
-  };
-
-
-
-  const GamesView = () => (
-    <div className="space-y-10">
-      <div className="border-b border-petal-rule pb-7">
-        <div className="flex flex-col md:flex-row md:items-end md:justify-between gap-6">
-          <div>
-            <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-              — 情趣
-            </div>
-            <h2 className="font-display text-4xl md:text-5xl font-light tracking-tight text-petal-ink leading-[1.05] mb-3">
-              情趣<em className="not-italic font-light italic text-pink-600">遊戲</em>
-            </h2>
-            <p className="font-display italic font-light text-base text-petal-muted">
-              增進彼此感情的遊戲、前戲與姿勢建議 — 一個個慢慢來。
-            </p>
-          </div>
-          <div className="font-display italic font-light text-xl text-petal-ink">
-            <Coins className="inline w-4 h-4 mr-1.5 text-petal-rose-deep" strokeWidth={1.5} />
-            <b className="not-italic font-medium">{totalCoins}</b> <span className="text-sm text-petal-muted">枚</span>
-          </div>
-        </div>
-      </div>
-
-      <nav className="!mt-0 sticky top-0 z-20 -mx-4 px-4 py-3 bg-petal-cream/95 backdrop-blur-sm border-b border-petal-rule-soft" aria-label="情趣遊戲分區">
-        <div className="overflow-x-auto no-scrollbar">
-          <div className="flex gap-2 whitespace-nowrap">
-            {[
-              { id: 'games-list', label: '情趣遊戲' },
-              { id: 'foreplay-activities', label: '前戲活動' },
-              { id: 'position-suggestions', label: '姿勢建議' },
-              { id: 'combo-suggestions', label: '連續組合技' },
-            ].map((section) => (
-              <button
-                key={section.id}
-                type="button"
-                onClick={() => {
-                  document.getElementById(section.id)?.scrollIntoView({ behavior: 'smooth', block: 'start' });
-                }}
-                className="px-3 py-1.5 border border-petal-rule rounded-full font-body text-xs text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink transition-colors"
-              >
-                {section.label}
-              </button>
-            ))}
-          </div>
-        </div>
-      </nav>
-
-      <div id="games-list" className="scroll-mt-24">
-        <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink mb-6">
-          情趣<em className="not-italic font-light italic text-pink-600">遊戲</em>
-        </h3>
-        <div className="space-y-6">
-        {romanticGames.map((game, index) => (
-          <div key={index} className="bg-white rounded-md border border-petal-rule p-7">
-            <div className="flex items-start space-x-4 mb-5 pb-5 border-b border-petal-rule-soft">
-              <div className="w-10 h-10 border border-petal-rose-soft bg-petal-rose-soft/40 rounded-full flex items-center justify-center">
-                <Sparkles className="w-4 h-4 text-petal-rose-deep" strokeWidth={1.5} />
-              </div>
-              <div className="flex-1">
-                <h3 className="font-display text-xl font-medium tracking-tight text-petal-ink mb-1">{game.title}</h3>
-                <p className="font-body text-sm text-petal-ink-soft leading-relaxed">{game.desc}</p>
-              </div>
-            </div>
-
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
-              <div className="bg-petal-cream-2/50 p-4 rounded-md">
-                <h4 className="font-body text-[11px] font-medium uppercase tracking-[0.14em] text-petal-muted mb-3">遊戲步驟</h4>
-                <ul className="space-y-2">
-                  {game.instructions.map((instruction, i) => (
-                    <li key={i} className="font-body text-sm text-petal-ink-soft leading-relaxed">{instruction}</li>
-                  ))}
-                </ul>
-              </div>
-
-              <div className="bg-petal-cream-2/50 p-4 rounded-md">
-                <h4 className="font-body text-[11px] font-medium uppercase tracking-[0.14em] text-petal-muted mb-3">
-                  {game.questions ? '問題範例' :
-                   game.tips ? '小貼士' :
-                   game.variations ? '變化玩法' :
-                   game.phrases ? '調情話語' :
-                   game.scenarios ? '場景建議' :
-                   game.emotions ? '情緒選項' : '願望類別'}
-                </h4>
-                <ul className="space-y-2">
-                  {(game.questions || game.tips || game.variations || game.phrases || game.scenarios || game.emotions || game.categories || []).map((item, i) => (
-                    <li key={i} className="font-body text-sm text-petal-ink-soft leading-relaxed">{item}</li>
-                  ))}
-                </ul>
-              </div>
-            </div>
-          </div>
-        ))}
-        </div>
-      </div>
-
-      <ForeplayView />
-    </div>
-  );
-
-  const ConflictView = () => {
-    const ArticleDivider = () => (
-      <div aria-hidden className="flex justify-center py-2">
-        <span className="font-display italic text-petal-muted/60 text-sm tracking-[0.5em]">· · ·</span>
-      </div>
-    );
-
-    // Pause mode — multi-step flow for couples already in a heated argument.
-    // Step 1: emotion selection · Step 2: safety phrase · Step 3: enforced
-    // turn-taking with 90s timer · Step 4: closing affirmation.
-    type EmotionKey = 'angry' | 'hurt' | 'cold' | 'overwhelmed';
-    const EMOTION_OPTIONS: { key: EmotionKey; emoji: string; label: string; sub: string; phrase: string }[] = [
-      {
-        key: 'angry',
-        emoji: '😡',
-        label: '生氣',
-        sub: '想反擊',
-        phrase: '我現在有點激動，我怕我會講出傷人的話，我想先停一下，但我不是不在乎你。',
-      },
-      {
-        key: 'hurt',
-        emoji: '😞',
-        label: '受傷',
-        sub: '想被理解',
-        phrase: '我現在其實有點難過，我需要你先聽我講完，不然我會更失落。',
-      },
-      {
-        key: 'cold',
-        emoji: '😐',
-        label: '冷掉',
-        sub: '不想講了',
-        phrase: '我現在腦袋有點關機，我需要一點時間消化，但我會回來，不是不在乎。',
-      },
-      {
-        key: 'overwhelmed',
-        emoji: '😣',
-        label: '快爆了',
-        sub: '撐不住了',
-        phrase: '我快撐不住了，我需要先暫停喘口氣，這不是不想處理，是先讓自己冷靜。',
-      },
-    ];
-
-    const TURN_SECONDS = 90;
-    const [pauseStep, setPauseStep] = useState<1 | 2 | 3 | 4 | null>(null);
-    useScrollLock(pauseStep !== null);
-    const [pauseEmotion, setPauseEmotion] = useState<EmotionKey | null>(null);
-    const [pauseRound, setPauseRound] = useState<1 | 2>(1);
-    const [pauseSeconds, setPauseSeconds] = useState(TURN_SECONDS);
-    const [listenerNudge, setListenerNudge] = useState<string | null>(null);
-
-    // Article expand/collapse — keep long-read tucked away by default so the page stays scannable.
-    const [articleExpanded, setArticleExpanded] = useState(false);
-
-    // Per-phrase send-to-partner state. Key = `${tier.key}-${phraseIndex}`.
-    const [phraseStatus, setPhraseStatus] = useState<Record<string, 'idle' | 'sending' | 'sent'>>({});
-
-    // Toolkit Step 2 accordion — which "need" is currently expanded.
-    const [expandedNeed, setExpandedNeed] = useState<string | null>(null);
-
-    // Custom-message composer state — per-step drafts and expanded/collapsed.
-    const [customDrafts, setCustomDrafts] = useState<Record<string, string>>({});
-    const [customExpanded, setCustomExpanded] = useState<Record<string, boolean>>({});
-
-    // Toolkit content — 4-step interactive flow for active conflict.
-    const emotionFullMessage = '我聽到你了，但我現在情緒有點滿，需要先停一下。我等一下會回來跟你說。';
-
-    const partnerNeedsItems: {
-      key: string;
-      icon: string;
-      title: string;
-      body: string;
-      suggestions?: string[];
-    }[] = [
-      {
-        key: 'pause-no-explain',
-        icon: '①',
-        title: '先停下來，不要解釋，不要辯論',
-        body: '我現在聽不進理性內容，你越解釋，我越痛。你只要先停下來，我會比較快冷靜。',
-      },
-      {
-        key: 'affirm-feeling',
-        icon: '②',
-        title: '用一句肯定我情緒的話（一定要有）',
-        body: '我不是要你道歉，是要你讓我覺得「你有看見我」。',
-        suggestions: [
-          '我知道我剛的話讓你不舒服了。',
-          '我看得出來你在生氣。',
-          '我先陪你一下。',
-        ],
-      },
-      {
-        key: 'give-step-down',
-        icon: '③',
-        title: '給我一個台階：你願意等我',
-        body: '有台階，我就會回頭。沒有台階，我只會越走越遠。',
-        suggestions: [
-          '你先休息一下，我在這裡等你。',
-          '等你準備好了，我們再一起聊。',
-        ],
-      },
-    ];
-
-    const exitOptions: { key: string; numeral: string; label: string; message: string }[] = [
-      {
-        key: 'pause-10min',
-        numeral: '1️⃣',
-        label: '暫停 10 分鐘',
-        message: '我會離開一下下，十分鐘後我們再回到同一個話題。',
-      },
-      {
-        key: 'pause-until-calm',
-        numeral: '2️⃣',
-        label: '暫停到情緒穩定',
-        message: '我現在想先讓情緒平穩，等我冷靜，我會主動來找你。',
-      },
-      {
-        key: 'switch-to-text',
-        numeral: '3️⃣',
-        label: '用文字聊代替現場聊',
-        message: '我現在用說的會越來越激動，我們改用訊息聊，好讓彼此舒服。',
-      },
-    ];
-
-    const tenderPhrases: string[] = [
-      '我沒有要跟你對立，我想靠近你。',
-      '你的感受我放在心上。',
-      '我在這裡，你慢慢來。',
-    ];
-
-    const soothingEmojis: string[] = ['🤝', '🫶', '🌙', '🤍', '🐻', '✨'];
-
-    const caringQuestion = '現在我能怎麼讓你舒服一點？';
-
-    const caringReplyOptions: string[] = [
-      '給我 5 分鐘就好。',
-      '說一句讓我安心的話。',
-      '先抱一下我。',
-    ];
-
-    const handleSendPhrase = async (phrase: string, key: string) => {
-      if (!partnerConnected) {
-        showNotification({
-          type: 'warning',
-          title: '尚未配對伴侶',
-          message: '需要先配對伴侶才能直接傳訊息給對方。',
-          duration: 4000,
-        });
-        return;
-      }
-      if (phraseStatus[key] === 'sending' || phraseStatus[key] === 'sent') return;
-      const confirmed = window.confirm(`確定要把這句話傳給TA嗎？\n\n「${phrase}」`);
-      if (!confirmed) return;
-      setPhraseStatus(prev => ({ ...prev, [key]: 'sending' }));
-      try {
-        await apiService.createIntimacyRequest({
-          messageContent: phrase,
-          requestType: 'reconciliation',
-        });
-        setPhraseStatus(prev => ({ ...prev, [key]: 'sent' }));
-        showNotification({
-          type: 'success',
-          title: '已送給TA',
-          message: '訊息已送到對方的邀請紀錄。',
-          duration: 3000,
-        });
-      } catch (err) {
-        setPhraseStatus(prev => ({ ...prev, [key]: 'idle' }));
-        showNotification({
-          type: 'error',
-          title: '送出失敗',
-          message: (err as Error)?.message || '請稍後再試。',
-          duration: 4000,
-        });
-      }
-    };
-
-    // Custom-message sender — same reconciliation channel, but uses the per-step textarea draft.
-    // Resets composer state on success so users can send multiple custom messages in a row.
-    const handleSendCustom = async (key: string) => {
-      const draft = (customDrafts[key] || '').trim();
-      if (!draft) {
-        showNotification({
-          type: 'warning',
-          title: '訊息不能為空',
-          message: '請先寫一些字再送出。',
-          duration: 3000,
-        });
-        return;
-      }
-      if (!partnerConnected) {
-        showNotification({
-          type: 'warning',
-          title: '尚未配對伴侶',
-          message: '需要先配對伴侶才能直接傳訊息給對方。',
-          duration: 4000,
-        });
-        return;
-      }
-      if (phraseStatus[key] === 'sending') return;
-      const confirmed = window.confirm(`確定要把這則自訂訊息傳給TA嗎？\n\n「${draft}」`);
-      if (!confirmed) return;
-      setPhraseStatus(prev => ({ ...prev, [key]: 'sending' }));
-      try {
-        await apiService.createIntimacyRequest({
-          messageContent: draft,
-          requestType: 'reconciliation',
-        });
-        showNotification({
-          type: 'success',
-          title: '已送給TA',
-          message: '自訂訊息已送到對方的邀請紀錄。',
-          duration: 3000,
-        });
-        setPhraseStatus(prev => ({ ...prev, [key]: 'idle' }));
-        setCustomDrafts(prev => ({ ...prev, [key]: '' }));
-        setCustomExpanded(prev => ({ ...prev, [key]: false }));
-      } catch (err) {
-        setPhraseStatus(prev => ({ ...prev, [key]: 'idle' }));
-        showNotification({
-          type: 'error',
-          title: '送出失敗',
-          message: (err as Error)?.message || '請稍後再試。',
-          duration: 4000,
-        });
-      }
-    };
-
-    // Inline custom-message composer — collapsed button by default, expands to textarea + send.
-    const renderCustomComposer = (key: string, placeholder = '寫一句自己想說的話傳給TA…') => {
-      const draft = customDrafts[key] || '';
-      const expanded = !!customExpanded[key];
-      const status = phraseStatus[key] || 'idle';
-      const isSending = status === 'sending';
-
-      if (!expanded) {
-        return (
-          <button
-            type="button"
-            onClick={() => setCustomExpanded(prev => ({ ...prev, [key]: true }))}
-            className="inline-flex items-center gap-1.5 px-3 py-1.5 rounded-full font-body text-xs font-medium border border-dashed border-petal-rule text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink hover:bg-white/70 transition-colors"
-          >
-            <Plus className="w-3.5 h-3.5" strokeWidth={1.75} />
-            寫自己的訊息
-          </button>
-        );
-      }
-
-      const trimmedLen = draft.trim().length;
-      const canSend = !isSending && trimmedLen > 0 && partnerConnected;
-
-      return (
-        <div className="bg-white/90 rounded-md border border-petal-rule p-3">
-          <textarea
-            value={draft}
-            onChange={(e) => setCustomDrafts(prev => ({ ...prev, [key]: e.target.value }))}
-            placeholder={placeholder}
-            rows={2}
-            maxLength={300}
-            disabled={isSending}
-            className="w-full font-body text-sm text-petal-ink leading-relaxed bg-transparent border-0 resize-none focus:outline-none placeholder:text-petal-muted/70"
-          />
-          <div className="flex items-center justify-between gap-2 mt-1.5 pt-2 border-t border-petal-rule-soft">
-            <span className="font-body text-[11px] text-petal-muted">
-              {draft.length}/300
-            </span>
-            <div className="flex items-center gap-2">
-              <button
-                type="button"
-                onClick={() => {
-                  setCustomExpanded(prev => ({ ...prev, [key]: false }));
-                  setCustomDrafts(prev => ({ ...prev, [key]: '' }));
-                }}
-                disabled={isSending}
-                className="px-3 py-1.5 rounded-full font-body text-xs text-petal-muted hover:text-petal-ink transition-colors"
-              >
-                取消
-              </button>
-              <button
-                type="button"
-                onClick={() => handleSendCustom(key)}
-                disabled={!canSend}
-                title={!partnerConnected ? '需要先配對伴侶' : ''}
-                className={`inline-flex items-center gap-1.5 px-3 py-1.5 rounded-full font-body text-xs font-medium border transition-colors ${
-                  isSending
-                    ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                    : canSend
-                      ? 'border-petal-rose bg-white text-petal-rose-deep hover:bg-petal-rose hover:text-white hover:border-petal-rose'
-                      : 'border-petal-rule bg-white text-petal-muted opacity-60 cursor-not-allowed'
-                }`}
-              >
-                {isSending ? (
-                  <>
-                    <span className="w-3.5 h-3.5 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                    傳送中
-                  </>
-                ) : (
-                  <>
-                    <Send className="w-3.5 h-3.5" strokeWidth={1.75} />
-                    傳給TA
-                  </>
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-      );
-    };
-
-    const scrollToSection = (id: string) => {
-      const el = document.getElementById(id);
-      if (!el) return;
-      el.scrollIntoView({ behavior: 'smooth', block: 'start' });
-    };
-
-    const sectionNav: { id: string; label: string }[] = [
-      { id: 'conflict-pause', label: '正在爭吵中' },
-      { id: 'conflict-toolkit', label: '應對工具' },
-      { id: 'conflict-article', label: '閱讀' },
-      { id: 'conflict-practice', label: '相處練習' },
-      { id: 'conflict-phrases', label: '給TA指引' },
-    ];
-
-    // Tick timer while on Step 3
-    useEffect(() => {
-      if (pauseStep !== 3) return;
-      const id = setInterval(() => setPauseSeconds(s => Math.max(0, s - 1)), 1000);
-      return () => clearInterval(id);
-    }, [pauseStep, pauseRound]);
-
-    // When timer hits 0, advance to next round or to Step 4
-    useEffect(() => {
-      if (pauseStep !== 3 || pauseSeconds > 0) return;
-      if (pauseRound === 1) {
-        setPauseRound(2);
-        setPauseSeconds(TURN_SECONDS);
-        setListenerNudge(null);
-      } else {
-        setPauseStep(4);
-      }
-    }, [pauseSeconds, pauseStep, pauseRound]);
-
-    const openPause = () => {
-      setPauseEmotion(null);
-      setPauseRound(1);
-      setPauseSeconds(TURN_SECONDS);
-      setListenerNudge(null);
-      setPauseStep(1);
-    };
-    const closePause = () => setPauseStep(null);
-
-    const pickEmotion = (key: EmotionKey) => {
-      setPauseEmotion(key);
-      setPauseStep(2);
-    };
-
-    const enterTurnTaking = () => {
-      setPauseRound(1);
-      setPauseSeconds(TURN_SECONDS);
-      setListenerNudge(null);
-      setPauseStep(3);
-    };
-
-    const skipTurn = () => {
-      if (pauseRound === 1) {
-        setPauseRound(2);
-        setPauseSeconds(TURN_SECONDS);
-        setListenerNudge(null);
-      } else {
-        setPauseStep(4);
-      }
-    };
-
-    const handleListenerReact = (label: string) => {
-      setListenerNudge(label);
-      window.setTimeout(() => {
-        setListenerNudge(current => (current === label ? null : current));
-      }, 2200);
-    };
-
-    const formatTime = (s: number) => {
-      const m = Math.floor(s / 60).toString().padStart(2, '0');
-      const ss = (s % 60).toString().padStart(2, '0');
-      return `${m}:${ss}`;
-    };
-
-    const selectedPhrase = pauseEmotion ? EMOTION_OPTIONS.find(o => o.key === pauseEmotion) : null;
-
-    return (
-    <div className="space-y-10">
-      <div className="border-b border-petal-rule pb-7">
-        <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-          — 和諧
-        </div>
-        <h2 className="font-display text-4xl md:text-5xl font-light tracking-tight text-petal-ink leading-[1.05] mb-3">
-          和諧<em className="not-italic font-light italic text-pink-600">相處</em>
-        </h2>
-        <p className="font-display italic font-light text-base text-petal-muted">
-          化解矛盾，增進理解 — 把急切的話留到明天再說。
-        </p>
-      </div>
-
-      {/* Sticky section nav — quick jumps within 和諧相處 */}
-      <nav
-        aria-label="頁面區塊"
-        className="sticky top-0 z-30 -mt-6 -mx-4 px-4 py-2.5 bg-petal-cream/95 backdrop-blur-sm border-b border-petal-rule"
-      >
-        <div className="flex gap-1.5 overflow-x-auto no-scrollbar -mx-1 px-1">
-          {sectionNav.map((s) => (
-            <button
-              key={s.id}
-              type="button"
-              onClick={() => scrollToSection(s.id)}
-              className="flex-shrink-0 px-3 py-1.5 rounded-full border border-petal-rule bg-white/70 text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink font-body text-xs font-medium tracking-tight transition-colors"
-            >
-              {s.label}
-            </button>
-          ))}
-        </div>
-      </nav>
-
-      {/* Emergency Pause — entry card for couples in active conflict */}
-      <div id="conflict-pause" className="bg-amber-50 border-2 border-amber-300 rounded-md p-5 md:p-6 scroll-mt-20">
-        <div className="flex flex-col md:flex-row md:items-center gap-4">
-          <div className="flex-1">
-            <h3 className="font-display text-lg font-medium tracking-tight text-amber-900 mb-1 flex items-center">
-              <Pause className="w-4 h-4 mr-2 fill-amber-700 text-amber-700" strokeWidth={1.5} />
-              正在爭吵中？
-            </h3>
-            <p className="font-body text-sm text-amber-900/80 leading-relaxed">
-              進入<b className="not-italic font-medium">冷靜連結模式</b>：4 步引導你們先降溫、再說話，而不是讓溝通變成武器。
-            </p>
-          </div>
-          <button
-            type="button"
-            onClick={openPause}
-            className="px-5 py-3 bg-amber-600 text-white rounded-md text-base font-medium hover:bg-amber-700 transition-colors whitespace-nowrap flex-shrink-0 flex items-center justify-center gap-2"
-          >
-            <Pause className="w-4 h-4 fill-white" strokeWidth={1.5} />
-            暫停一下
-          </button>
-        </div>
-      </div>
-
-      {/* Both-sides conflict toolkit — 4-step interactive guide with sendable messages per step */}
-      <section id="conflict-toolkit" className="scroll-mt-20">
-        <header className="mb-6">
-          <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-            — 衝突進行中的工具箱
-          </div>
-          <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink mb-2">
-            雙方衝突時的<em className="not-italic font-light italic text-pink-600">應對工具</em>
-          </h3>
-          <p className="font-body text-sm text-petal-ink-soft leading-relaxed">
-            四個小步驟，從「我滿了」走到「我們重新靠近」。每一步都有一鍵就能送到TA那裡的訊息。
-            {!partnerConnected && (
-              <span className="block mt-1.5 text-petal-muted text-xs">
-                配對伴侶後，點任何「傳給TA」按鈕即可直接送出。
-              </span>
-            )}
-          </p>
-        </header>
-
-        <div className="space-y-5">
-          {/* Step 1 — Notify TA that you're emotionally full */}
-          <article className="bg-violet-50/60 border-2 border-violet-200 rounded-md p-5 md:p-7">
-            <header className="mb-3 flex flex-wrap items-baseline gap-x-3 gap-y-2">
-              <span className="inline-flex items-center gap-1.5 font-body text-[11px] font-semibold uppercase tracking-[0.16em] text-violet-800 bg-violet-100/80 px-2.5 py-1 rounded">
-                <span aria-hidden>🟣</span>
-                Step 1
-              </span>
-              <h4 className="font-display text-lg md:text-xl font-medium text-petal-ink leading-snug">
-                我現在情緒滿了
-              </h4>
-            </header>
-            <p className="font-body text-[13px] text-petal-ink-soft mb-4 leading-relaxed">
-              用一句話讓TA知道你需要空間 — 不是生氣、不是不想聽，只是情緒太滿了。
-            </p>
-            <blockquote className="font-display italic text-[15px] text-petal-ink bg-white/70 rounded-md p-4 border border-white mb-4 leading-relaxed">
-              「我現在需要一下下的空間，不是生你的氣，也不是不想聽，只是情緒太滿了。」
-            </blockquote>
-
-            <div className="bg-white/80 rounded-md p-4 border border-violet-200/50">
-              <div className="font-body text-[11px] uppercase tracking-[0.14em] text-violet-800/80 mb-2">
-                📩 一鍵傳送（自動訊息）
-              </div>
-              <p className="font-display italic text-[14px] text-petal-ink mb-3 leading-relaxed">
-                「{emotionFullMessage}」
-              </p>
-              {(() => {
-                const key = 'toolkit-step1-notify';
-                const status = phraseStatus[key] || 'idle';
-                return (
-                  <button
-                    type="button"
-                    onClick={() => handleSendPhrase(emotionFullMessage, key)}
-                    disabled={status === 'sending' || status === 'sent'}
-                    className={`inline-flex items-center gap-2 px-4 py-2.5 rounded-md font-body text-sm font-medium border transition-colors ${
-                      status === 'sent'
-                        ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                        : status === 'sending'
-                          ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                          : partnerConnected
-                            ? 'border-violet-400 bg-violet-600 text-white hover:bg-violet-700 hover:border-violet-700'
-                            : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                    }`}
-                  >
-                    {status === 'sent' ? (
-                      <>
-                        <Check className="w-4 h-4" strokeWidth={2} />
-                        已傳送：情緒已滿的通知
-                      </>
-                    ) : status === 'sending' ? (
-                      <>
-                        <span className="w-4 h-4 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                        傳送中…
-                      </>
-                    ) : (
-                      <>
-                        <Send className="w-4 h-4" strokeWidth={1.75} />
-                        傳送：情緒已滿的通知
-                      </>
-                    )}
-                  </button>
-                );
-              })()}
-            </div>
-
-            <div className="mt-4">
-              {renderCustomComposer('toolkit-step1-custom', '想用自己的話告訴TA「我滿了」？寫在這…')}
-            </div>
-          </article>
-
-          {/* Step 2 — 3 things TA most needs to know (accordion) */}
-          <article className="bg-sky-50/60 border-2 border-sky-200 rounded-md p-5 md:p-7">
-            <header className="mb-3 flex flex-wrap items-baseline gap-x-3 gap-y-2">
-              <span className="inline-flex items-center gap-1.5 font-body text-[11px] font-semibold uppercase tracking-[0.16em] text-sky-800 bg-sky-100/80 px-2.5 py-1 rounded">
-                <span aria-hidden>🔵</span>
-                Step 2
-              </span>
-              <h4 className="font-display text-lg md:text-xl font-medium text-petal-ink leading-snug">
-                給TA看的：3 件我最需要的事情
-              </h4>
-            </header>
-            <p className="font-body text-[13px] text-petal-ink-soft mb-4 leading-relaxed">
-              點開每一項，讓TA了解你最需要什麼。展開後的建議話語可以一鍵傳給TA。
-            </p>
-
-            <div className="space-y-2">
-              {partnerNeedsItems.map((item) => {
-                const isExpanded = expandedNeed === item.key;
-                return (
-                  <div key={item.key} className="bg-white/80 rounded-md border border-sky-200/50 overflow-hidden">
-                    <button
-                      type="button"
-                      onClick={() => setExpandedNeed(isExpanded ? null : item.key)}
-                      className="w-full px-4 py-3 flex items-center gap-3 text-left hover:bg-sky-50/40 transition-colors"
-                      aria-expanded={isExpanded}
-                    >
-                      <span className="font-display text-petal-rose-deep text-base flex-shrink-0">{item.icon}</span>
-                      <span className="font-display text-[15px] font-medium text-petal-ink flex-1 leading-snug">
-                        {item.title}
-                      </span>
-                      <ChevronDown
-                        className={`w-4 h-4 text-petal-muted flex-shrink-0 transform transition-transform ${isExpanded ? 'rotate-180' : ''}`}
-                        strokeWidth={1.5}
-                      />
-                    </button>
-                    {isExpanded && (
-                      <div className="px-4 pb-4 pt-1 border-t border-sky-100">
-                        <p className="font-body text-[13px] text-petal-ink-soft mb-3 leading-relaxed">
-                          {item.body}
-                        </p>
-                        {item.suggestions && (
-                          <>
-                            <div className="font-body text-[11px] uppercase tracking-[0.14em] text-sky-800/80 mb-2">
-                              建議話語
-                            </div>
-                            <ul className="space-y-2">
-                              {item.suggestions.map((s, i) => {
-                                const key = `toolkit-step2-${item.key}-${i}`;
-                                const status = phraseStatus[key] || 'idle';
-                                return (
-                                  <li
-                                    key={i}
-                                    className="flex flex-col sm:flex-row sm:items-start gap-2 bg-white rounded-md p-3 border border-sky-100"
-                                  >
-                                    <blockquote className="font-display text-[14px] text-petal-ink leading-relaxed flex-1">
-                                      「{s}」
-                                    </blockquote>
-                                    <button
-                                      type="button"
-                                      onClick={() => handleSendPhrase(s, key)}
-                                      disabled={status === 'sending' || status === 'sent'}
-                                      title={partnerConnected ? '傳送這句話給TA' : '需要先配對伴侶'}
-                                      className={`flex-shrink-0 self-end sm:self-start inline-flex items-center gap-1.5 px-3 py-1.5 rounded-full font-body text-xs font-medium border transition-colors ${
-                                        status === 'sent'
-                                          ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                                          : status === 'sending'
-                                            ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                                            : partnerConnected
-                                              ? 'border-sky-300 bg-white text-sky-800 hover:bg-sky-600 hover:text-white hover:border-sky-600'
-                                              : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                                      }`}
-                                    >
-                                      {status === 'sent' ? (
-                                        <>
-                                          <Check className="w-3.5 h-3.5" strokeWidth={2} />
-                                          已送出
-                                        </>
-                                      ) : status === 'sending' ? (
-                                        <>
-                                          <span className="w-3.5 h-3.5 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                                          傳送中
-                                        </>
-                                      ) : (
-                                        <>
-                                          <Send className="w-3.5 h-3.5" strokeWidth={1.75} />
-                                          傳給TA
-                                        </>
-                                      )}
-                                    </button>
-                                  </li>
-                                );
-                              })}
-                            </ul>
-                          </>
-                        )}
-                      </div>
-                    )}
-                  </div>
-                );
-              })}
-            </div>
-
-            <div className="mt-4">
-              {renderCustomComposer('toolkit-step2-custom', '想自己跟TA說「我最需要的是…」？寫在這…')}
-            </div>
-          </article>
-
-          {/* Step 3 — 3 exit options for pausing the conflict */}
-          <article className="bg-orange-50/60 border-2 border-orange-200 rounded-md p-5 md:p-7">
-            <header className="mb-3 flex flex-wrap items-baseline gap-x-3 gap-y-2">
-              <span className="inline-flex items-center gap-1.5 font-body text-[11px] font-semibold uppercase tracking-[0.16em] text-orange-800 bg-orange-100/80 px-2.5 py-1 rounded">
-                <span aria-hidden>🟠</span>
-                Step 3
-              </span>
-              <h4 className="font-display text-lg md:text-xl font-medium text-petal-ink leading-snug">
-                衝突升起時的退場流程
-              </h4>
-            </header>
-            <p className="font-body text-[13px] text-petal-ink-soft mb-4 leading-relaxed">
-              選一種暫停方式，一鍵把對應的暫停訊息送給TA。
-            </p>
-
-            <div className="space-y-3">
-              {exitOptions.map((opt) => {
-                const key = `toolkit-step3-${opt.key}`;
-                const status = phraseStatus[key] || 'idle';
-                return (
-                  <div key={opt.key} className="bg-white/80 rounded-md p-4 border border-orange-200/50">
-                    <div className="flex items-center gap-2 mb-2">
-                      <span className="text-base" aria-hidden>{opt.numeral}</span>
-                      <span className="font-display text-[15px] font-medium text-petal-ink">
-                        {opt.label}
-                      </span>
-                    </div>
-                    <p className="font-display italic text-[14px] text-petal-ink-soft mb-3 leading-relaxed">
-                      「{opt.message}」
-                    </p>
-                    <button
-                      type="button"
-                      onClick={() => handleSendPhrase(opt.message, key)}
-                      disabled={status === 'sending' || status === 'sent'}
-                      title={partnerConnected ? '傳送這個暫停訊息給TA' : '需要先配對伴侶'}
-                      className={`inline-flex items-center gap-2 px-4 py-2 rounded-md font-body text-sm font-medium border transition-colors ${
-                        status === 'sent'
-                          ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                          : status === 'sending'
-                            ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                            : partnerConnected
-                              ? 'border-orange-400 bg-white text-orange-800 hover:bg-orange-600 hover:text-white hover:border-orange-600'
-                              : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                      }`}
-                    >
-                      {status === 'sent' ? (
-                        <>
-                          <Check className="w-4 h-4" strokeWidth={2} />
-                          已傳送
-                        </>
-                      ) : status === 'sending' ? (
-                        <>
-                          <span className="w-4 h-4 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                          傳送中…
-                        </>
-                      ) : (
-                        <>
-                          <Send className="w-4 h-4" strokeWidth={1.75} />
-                          選這個 — 傳給TA
-                        </>
-                      )}
-                    </button>
-                  </div>
-                );
-              })}
-            </div>
-
-            <div className="mt-4">
-              {renderCustomComposer('toolkit-step3-custom', '想自己寫一個暫停的方式？例如「等我洗完澡再聊」…')}
-            </div>
-          </article>
-
-          {/* Step 4 — 3 small tasks to bring me back */}
-          <article className="bg-amber-50/70 border-2 border-amber-200 rounded-md p-5 md:p-7">
-            <header className="mb-3 flex flex-wrap items-baseline gap-x-3 gap-y-2">
-              <span className="inline-flex items-center gap-1.5 font-body text-[11px] font-semibold uppercase tracking-[0.16em] text-amber-900 bg-amber-100/80 px-2.5 py-1 rounded">
-                <span aria-hidden>🟡</span>
-                Step 4
-              </span>
-              <h4 className="font-display text-lg md:text-xl font-medium text-petal-ink leading-snug">
-                讓我冷靜後願意回來的關鍵
-              </h4>
-            </header>
-            <p className="font-body text-[13px] text-petal-ink-soft mb-5 leading-relaxed">
-              當你準備好靠近TA時 — 或想告訴TA「這樣對我最有效」 — 這三件小事最能拉近距離。
-            </p>
-
-            {/* Task 1 — tender phrase */}
-            <div className="mb-6">
-              <div className="font-body text-[12px] font-semibold tracking-tight text-amber-900 mb-2 flex items-center gap-1.5">
-                <span aria-hidden>✔</span>
-                任務 1 — 傳一句溫柔語氣
-              </div>
-              <ul className="space-y-2">
-                {tenderPhrases.map((p, i) => {
-                  const key = `toolkit-step4-tender-${i}`;
-                  const status = phraseStatus[key] || 'idle';
-                  return (
-                    <li
-                      key={i}
-                      className="flex flex-col sm:flex-row sm:items-start gap-2 bg-white/80 rounded-md p-3 border border-amber-200/50"
-                    >
-                      <blockquote className="font-display text-[14px] text-petal-ink leading-relaxed flex-1">
-                        「{p}」
-                      </blockquote>
-                      <button
-                        type="button"
-                        onClick={() => handleSendPhrase(p, key)}
-                        disabled={status === 'sending' || status === 'sent'}
-                        title={partnerConnected ? '傳送這句話給TA' : '需要先配對伴侶'}
-                        className={`flex-shrink-0 self-end sm:self-start inline-flex items-center gap-1.5 px-3 py-1.5 rounded-full font-body text-xs font-medium border transition-colors ${
-                          status === 'sent'
-                            ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                            : status === 'sending'
-                              ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                              : partnerConnected
-                                ? 'border-amber-400 bg-white text-amber-800 hover:bg-amber-600 hover:text-white hover:border-amber-600'
-                                : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                        }`}
-                      >
-                        {status === 'sent' ? (
-                          <>
-                            <Check className="w-3.5 h-3.5" strokeWidth={2} />
-                            已送出
-                          </>
-                        ) : status === 'sending' ? (
-                          <>
-                            <span className="w-3.5 h-3.5 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                            傳送中
-                          </>
-                        ) : (
-                          <>
-                            <Send className="w-3.5 h-3.5" strokeWidth={1.75} />
-                            傳給TA
-                          </>
-                        )}
-                      </button>
-                    </li>
-                  );
-                })}
-              </ul>
-            </div>
-
-            {/* Task 2 — soothing emoji */}
-            <div className="mb-6">
-              <div className="font-body text-[12px] font-semibold tracking-tight text-amber-900 mb-2 flex items-center gap-1.5">
-                <span aria-hidden>✔</span>
-                任務 2 — 傳一個安撫 emoji
-              </div>
-              <p className="font-body text-xs text-petal-ink-soft mb-3 leading-relaxed">
-                點一個 emoji 就送出，不用任何文字。
-              </p>
-              <div className="flex flex-wrap gap-2">
-                {soothingEmojis.map((emoji, i) => {
-                  const key = `toolkit-step4-emoji-${i}`;
-                  const status = phraseStatus[key] || 'idle';
-                  const isSent = status === 'sent';
-                  const isSending = status === 'sending';
-                  return (
-                    <button
-                      key={i}
-                      type="button"
-                      onClick={() => handleSendPhrase(emoji, key)}
-                      disabled={isSent || isSending}
-                      title={
-                        isSent ? '已送出' : partnerConnected ? `傳送 ${emoji} 給TA` : '需要先配對伴侶'
-                      }
-                      className={`w-12 h-12 text-2xl rounded-full border-2 flex items-center justify-center transition-colors ${
-                        isSent
-                          ? 'border-petal-sage bg-petal-sage/15 cursor-default'
-                          : isSending
-                            ? 'border-petal-rule bg-white opacity-60 cursor-wait'
-                            : partnerConnected
-                              ? 'border-amber-300 bg-white hover:bg-amber-100 hover:border-amber-500'
-                              : 'border-petal-rule bg-white opacity-60'
-                      }`}
-                    >
-                      {emoji}
-                    </button>
-                  );
-                })}
-              </div>
-            </div>
-
-            {/* Task 3 — caring question with pre-prepared replies */}
-            <div>
-              <div className="font-body text-[12px] font-semibold tracking-tight text-amber-900 mb-2 flex items-center gap-1.5">
-                <span aria-hidden>✔</span>
-                任務 3 — 問一個貼心小問題
-              </div>
-              <div className="bg-white/80 rounded-md p-4 border border-amber-200/50">
-                <blockquote className="font-display italic text-[15px] text-petal-ink mb-3 leading-relaxed">
-                  「{caringQuestion}」
-                </blockquote>
-                {(() => {
-                  const key = 'toolkit-step4-question';
-                  const status = phraseStatus[key] || 'idle';
-                  return (
-                    <button
-                      type="button"
-                      onClick={() => handleSendPhrase(caringQuestion, key)}
-                      disabled={status === 'sending' || status === 'sent'}
-                      title={partnerConnected ? '把這個問題傳給TA' : '需要先配對伴侶'}
-                      className={`inline-flex items-center gap-2 px-4 py-2 rounded-md font-body text-sm font-medium border transition-colors ${
-                        status === 'sent'
-                          ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                          : status === 'sending'
-                            ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                            : partnerConnected
-                              ? 'border-amber-400 bg-white text-amber-800 hover:bg-amber-600 hover:text-white hover:border-amber-600'
-                              : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                      }`}
-                    >
-                      {status === 'sent' ? (
-                        <>
-                          <Check className="w-4 h-4" strokeWidth={2} />
-                          已傳送
-                        </>
-                      ) : status === 'sending' ? (
-                        <>
-                          <span className="w-4 h-4 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                          傳送中…
-                        </>
-                      ) : (
-                        <>
-                          <Send className="w-4 h-4" strokeWidth={1.75} />
-                          傳這句話給TA
-                        </>
-                      )}
-                    </button>
-                  );
-                })()}
-
-                <div className="font-body text-[11px] uppercase tracking-[0.14em] text-petal-muted mt-5 mb-2">
-                  你可以選的回覆（一鍵送出）
-                </div>
-                <ul className="space-y-2">
-                  {caringReplyOptions.map((r, i) => {
-                    const key = `toolkit-step4-reply-${i}`;
-                    const status = phraseStatus[key] || 'idle';
-                    return (
-                      <li
-                        key={i}
-                        className="flex flex-col sm:flex-row sm:items-start gap-2 bg-white rounded-md p-3 border border-amber-100"
-                      >
-                        <blockquote className="font-display text-[14px] text-petal-ink leading-relaxed flex-1">
-                          「{r}」
-                        </blockquote>
-                        <button
-                          type="button"
-                          onClick={() => handleSendPhrase(r, key)}
-                          disabled={status === 'sending' || status === 'sent'}
-                          title={partnerConnected ? '把這個回覆送給TA' : '需要先配對伴侶'}
-                          className={`flex-shrink-0 self-end sm:self-start inline-flex items-center gap-1.5 px-3 py-1.5 rounded-full font-body text-xs font-medium border transition-colors ${
-                            status === 'sent'
-                              ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                              : status === 'sending'
-                                ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                                : partnerConnected
-                                  ? 'border-amber-400 bg-white text-amber-800 hover:bg-amber-600 hover:text-white hover:border-amber-600'
-                                  : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                          }`}
-                        >
-                          {status === 'sent' ? (
-                            <>
-                              <Check className="w-3.5 h-3.5" strokeWidth={2} />
-                              已送出
-                            </>
-                          ) : status === 'sending' ? (
-                            <>
-                              <span className="w-3.5 h-3.5 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                              傳送中
-                            </>
-                          ) : (
-                            <>
-                              <Send className="w-3.5 h-3.5" strokeWidth={1.75} />
-                              傳給TA
-                            </>
-                          )}
-                        </button>
-                      </li>
-                    );
-                  })}
-                </ul>
-              </div>
-            </div>
-
-            <div className="mt-5 pt-5 border-t border-amber-200/60">
-              {renderCustomComposer('toolkit-step4-custom', '想自己寫一句靠近TA的話？寫在這…')}
-            </div>
-          </article>
-        </div>
-
-        <p className="font-display italic text-center text-sm text-petal-muted mt-6 leading-relaxed">
-          工具是橋 — 一步步走，回到彼此身邊。
-        </p>
-      </section>
-
-      {/* Featured Long-Read — 翻譯彼此的語言 (collapsible to keep page scannable) */}
-      <article id="conflict-article" className="bg-white rounded-md border border-petal-rule p-6 md:p-10 scroll-mt-20">
-        <header className={articleExpanded ? 'mb-7 pb-5 border-b border-petal-rule-soft' : ''}>
-          <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-            — 閱讀
-          </div>
-          <h3 className="font-display text-3xl font-light tracking-tight text-petal-ink leading-[1.15] mb-3">
-            翻譯彼此的<em className="not-italic font-light italic text-pink-600">語言</em>
-          </h3>
-          <p className="font-display italic font-light text-sm text-petal-muted leading-relaxed">
-            在親密關係裡，最常見的衝突往往不是「不愛了」，而是「聽不懂彼此在說什麼」。
-          </p>
-        </header>
-
-        {!articleExpanded && (
-          <div className="mt-5 pt-5 border-t border-petal-rule-soft flex flex-col sm:flex-row sm:items-center sm:justify-between gap-3">
-            <p className="font-body text-xs text-petal-muted leading-relaxed">
-              一則關於先生與太太、靠近與界線的長文 — 約 5 分鐘閱讀。
-            </p>
-            <button
-              type="button"
-              onClick={() => setArticleExpanded(true)}
-              className="inline-flex items-center justify-center gap-1.5 px-4 py-2 border border-petal-rule rounded-md font-body text-sm text-petal-rose-deep hover:border-petal-rose-deep hover:bg-petal-cream-2/40 transition-colors self-start sm:self-auto"
-            >
-              閱讀全文
-              <ChevronDown className="w-4 h-4" strokeWidth={1.5} />
-            </button>
-          </div>
-        )}
-
-        {articleExpanded && (
-        <div className="space-y-5 font-body text-[15px] leading-loose text-petal-ink">
-          <p>在一段親密關係裡，最常見的衝突，往往不是「不愛了」，而是「聽不懂彼此在說什麼」。</p>
-          <p>很多時候，兩個人其實都還在乎對方，只是用完全不同的方式在表達需求。</p>
-
-          <ArticleDivider />
-
-          <p>有一對伴侶，先生和太太，正在經歷這樣的狀況。</p>
-          <p>先生其實不是只想要性。他更多時候，是在某些疲憊或空虛的瞬間，想靠近太太，確認彼此還是連結著的。他會想抱一下、想靠近一點、想讓對方知道：「我還在這裡，我想你也還在我身邊。」</p>
-          <p>而太太其實也不是不在乎他。只是她常常在忙碌、疲累，或情緒沒有空間的時候，需要安靜與界線。</p>
-          <p>問題是，兩個人的「靠近方式」開始產生錯位。</p>
-
-          <ArticleDivider />
-
-          <p>有一次，先生又在晚上試著靠近她。他沒有很強烈地要求什麼，只是用那種想要親近的方式，想拉近距離。</p>
-          <p>但太太那天已經很累了。她一開始其實是溫柔的，她說：「我現在有點累。」</p>
-          <p>但先生還是想再靠近一點、再確認一點。</p>
-          <p>不是故意的，只是他沒有接收到那個「已經不行了」的訊號已經很清楚。</p>
-
-          <ArticleDivider />
-
-          <p>太太開始感覺到一種熟悉的壓力——她明明已經很溫柔了，但好像還是不被停下來。</p>
-          <p>於是她的語氣變了。</p>
-          <p>從溫柔，變成比較直接。</p>
-          <p>甚至有一點冷。</p>
-          <p>她心裡想的是：「如果我不講得更清楚，你是不是不會停？」</p>
-          <p>但她沒有說出口的是——她並不是不想連結，而是她已經沒有力氣再用更溫柔的方式了。</p>
-
-          <ArticleDivider />
-
-          <p>先生在那一刻感受到的，是另一種痛。</p>
-          <p>他不是被「需求」拒絕，而是感覺到自己好像被推開了。</p>
-          <p>他心裡會想：「我只是想靠近你，為什麼你變得這麼冷？」</p>
-          <p>但他沒有看見的是，太太的強硬，其實不是遠離，而是一種防禦。</p>
-          <p>她不是在說「我不愛你」，而是在說：「我需要你現在先停一下，因為我已經沒有空間了。」</p>
-
-          <ArticleDivider />
-
-          <p>這就是兩人之間最常見的誤會。</p>
-          <p className="font-display italic text-petal-ink-soft">當溫柔沒有被理解時，會變成更強的界線。</p>
-          <p className="font-display italic text-petal-ink-soft">當界線被感覺成拒絕時，會變成更深的失落。</p>
-          <p>兩個人都在保護自己，但方式卻讓彼此越來越遠。</p>
-
-          <ArticleDivider />
-
-          <p>後來，他們開始學著翻譯彼此的語言。</p>
-          <p>太太慢慢發現，先生很多時候要的不是「做什麼」，而是「還連不連得上」。</p>
-          <p>於是她開始嘗試，在無法配合的時候，不只是說「不行」，而是補上一點連結：</p>
-          <ul className="space-y-2 pl-5 list-disc marker:text-petal-rose-deep">
-            <li>輕輕抱一下</li>
-            <li>摸一下手</li>
-            <li>說一句：「我現在真的累，但我還在」</li>
-            <li>或是給一個明確時間：「等一下／晚點／明天」</li>
-          </ul>
-          <p>這些小小的動作，對先生來說，比任何解釋都更有安定感。</p>
-
-          <ArticleDivider />
-
-          <p>而先生也開始理解，太太的直接，不是冷漠，而是她的極限已經到了。</p>
-          <p>她不是不想溫柔，而是她曾經溫柔過，但那種溫柔沒有被理解，於是她只能換一種更清楚的方式保護自己。</p>
-
-          <ArticleDivider />
-
-          <p>慢慢地，他們才發現一件很重要的事：</p>
-          <p className="font-display italic text-lg text-petal-ink leading-relaxed">親密關係的問題，從來不是「要不要答應」，而是「有沒有被理解」。</p>
-
-          <ArticleDivider />
-
-          <div className="space-y-3">
-            <p><span className="font-medium text-petal-rose-deep">男生</span>想要的，很多時候不是結果，而是被肯定、被看見、被連結。</p>
-            <p><span className="font-medium text-petal-rose-deep">女生</span>需要的，很多時候不是拒絕，而是界線被尊重，同時關係沒有斷掉。</p>
-          </div>
-
-          <ArticleDivider />
-
-          <p>當兩個人開始願意翻譯彼此的語言時，關係會開始改變。</p>
-          <p>不再是「你要或不要」。</p>
-          <p>而變成：</p>
-
-          <blockquote className="my-2 py-4 px-5 border-l-2 border-petal-rose-soft bg-petal-cream-2/40 font-display italic text-base leading-relaxed text-petal-ink space-y-2">
-            <p>「我現在不行，但我還在。」</p>
-            <p>「我聽見你了，但我們換一種方式靠近。」</p>
-          </blockquote>
-
-          <ArticleDivider />
-
-          <p className="font-display italic text-lg leading-relaxed text-petal-ink text-center py-2">
-            親密關係真正的成熟，不是永遠一致，而是即使不同步，也不讓彼此失聯。
-          </p>
-
-          <div className="pt-5 border-t border-petal-rule-soft flex justify-center">
-            <button
-              type="button"
-              onClick={() => {
-                setArticleExpanded(false);
-                scrollToSection('conflict-article');
-              }}
-              className="inline-flex items-center gap-1.5 px-4 py-2 border border-petal-rule rounded-md font-body text-sm text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink transition-colors"
-            >
-              <ChevronUp className="w-4 h-4" strokeWidth={1.5} />
-              收起
-            </button>
-          </div>
-        </div>
-        )}
-      </article>
-
-      <div id="conflict-practice" className="scroll-mt-20">
-        <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink mb-6">
-          相處<em className="not-italic font-light italic text-pink-600">練習</em>
-        </h3>
-        <div className="space-y-4">
-          {conflictResolutions.map((solution, index) => (
-            <div key={index} className="bg-white rounded-md border border-petal-rule p-6">
-              <div className="flex items-start space-x-4">
-                <div className="w-10 h-10 border border-petal-sage/60 bg-petal-sage/10 rounded-full flex items-center justify-center">
-                  <MessageCircle className="w-4 h-4 text-petal-sage-deep" strokeWidth={1.5} />
-                </div>
-                <div className="flex-1">
-                  <h3 className="font-display text-lg font-medium tracking-tight text-petal-ink mb-1.5">{solution.title}</h3>
-                  <p className="font-body text-sm text-petal-ink-soft leading-relaxed">{solution.desc}</p>
-                </div>
-              </div>
-            </div>
-          ))}
-        </div>
-      </div>
-
-      {/* Partner-facing guide — what TA can do when I'm upset, with sendable phrases per step */}
-      <section id="conflict-phrases" className="scroll-mt-20">
-        <header className="mb-6">
-          <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-            — 給TA的對話指引
-          </div>
-          <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink mb-2">
-            當我生氣時，你可以<em className="not-italic font-light italic text-pink-600">這樣做</em>
-          </h3>
-          <p className="font-body text-sm text-petal-ink-soft leading-relaxed">
-            八個小步驟，不是要你道歉或認錯 — 而是希望我們在情緒裡也能靠近。你越溫柔，我越能下台階。
-            {partnerConnected && (
-              <span className="block mt-1.5 text-petal-muted text-xs">
-                點每一句後面的「傳給TA」，就能把這句話直接送到對方的邀請紀錄。
-              </span>
-            )}
-          </p>
-        </header>
-
-        <div className="space-y-5">
-          {conflictPhraseTiers.map((tier) => (
-            <article
-              key={tier.key}
-              className={`rounded-md border-2 ${tier.cardClass} p-5 md:p-7`}
-            >
-              <header className="mb-4 flex flex-wrap items-baseline gap-x-3 gap-y-2">
-                <span
-                  className={`inline-flex items-center gap-1.5 font-body text-[11px] font-semibold uppercase tracking-[0.16em] ${tier.badgeClass} px-2.5 py-1 rounded`}
-                >
-                  <span aria-hidden>{tier.dot}</span>
-                  {tier.badge}
-                </span>
-                <h4 className="font-display text-lg md:text-xl font-medium text-petal-ink leading-snug">
-                  {tier.title}
-                </h4>
-              </header>
-
-              <p className="font-body text-[13px] text-petal-ink-soft mb-5 leading-relaxed">
-                {tier.why}
-              </p>
-
-              <ol className="space-y-3 mb-5">
-                {tier.phrases.map((phrase, i) => {
-                  const phraseKey = `${tier.key}-${i}`;
-                  const status = phraseStatus[phraseKey] || 'idle';
-                  return (
-                    <li
-                      key={i}
-                      className="flex flex-col sm:flex-row sm:items-start gap-3 bg-white/70 rounded-md p-3.5 md:p-4 border border-white"
-                    >
-                      <div className="flex gap-3 flex-1 min-w-0">
-                        <span className="font-display italic text-petal-rose-deep text-sm leading-relaxed flex-shrink-0 mt-0.5 w-5">
-                          {i + 1}.
-                        </span>
-                        <blockquote className="font-display text-[15px] md:text-base leading-relaxed text-petal-ink flex-1">
-                          「{phrase}」
-                        </blockquote>
-                      </div>
-                      <button
-                        type="button"
-                        onClick={() => handleSendPhrase(phrase, phraseKey)}
-                        disabled={status === 'sending' || status === 'sent'}
-                        title={partnerConnected ? '直接傳送這句話給TA' : '需要先配對伴侶'}
-                        className={`flex-shrink-0 self-end sm:self-start inline-flex items-center gap-1.5 px-3 py-1.5 rounded-full font-body text-xs font-medium tracking-tight border transition-colors ${
-                          status === 'sent'
-                            ? 'border-petal-sage bg-petal-sage/15 text-petal-sage-deep cursor-default'
-                            : status === 'sending'
-                              ? 'border-petal-rule bg-white text-petal-muted cursor-wait'
-                              : partnerConnected
-                                ? 'border-petal-rose bg-white text-petal-rose-deep hover:bg-petal-rose hover:text-white hover:border-petal-rose'
-                                : 'border-petal-rule bg-white text-petal-muted opacity-60'
-                        }`}
-                      >
-                        {status === 'sent' ? (
-                          <>
-                            <Check className="w-3.5 h-3.5" strokeWidth={2} />
-                            已送出
-                          </>
-                        ) : status === 'sending' ? (
-                          <>
-                            <span className="w-3.5 h-3.5 border-2 border-petal-muted border-t-transparent rounded-full animate-spin" />
-                            傳送中
-                          </>
-                        ) : (
-                          <>
-                            <Send className="w-3.5 h-3.5" strokeWidth={1.75} />
-                            傳給TA
-                          </>
-                        )}
-                      </button>
-                    </li>
-                  );
-                })}
-              </ol>
-
-              <div className="mb-5">
-                {renderCustomComposer(`${tier.key}-custom`, `這一步想自己說一句話？寫給TA…`)}
-              </div>
-
-              {tier.note && (
-                <p className="font-body text-xs md:text-sm text-petal-ink-soft leading-relaxed border-t border-petal-rule-soft pt-3">
-                  {tier.note}
-                </p>
-              )}
-            </article>
-          ))}
-        </div>
-
-        <p className="font-display italic text-center text-sm text-petal-muted mt-6 leading-relaxed">
-          你越溫柔，我越能回來 — 不是輸贏，是回到彼此身邊。
-        </p>
-      </section>
-
-      {/* Pause Mode — full-screen guided flow */}
-      {pauseStep !== null && (
-        <div className="fixed inset-0 z-50 bg-petal-cream overflow-y-auto">
-          <div className="max-w-2xl mx-auto px-5 sm:px-8 py-8 min-h-screen flex flex-col">
-            {/* Top bar */}
-            <div className="flex items-start justify-between mb-8">
-              <div>
-                <div className="font-body text-[11px] uppercase tracking-[0.18em] text-amber-700 mb-1">— 暫停模式</div>
-                <div className="font-display text-xl text-petal-ink">冷靜連結模式</div>
-                <div className="font-body text-xs text-petal-muted mt-0.5">Step {pauseStep} / 4</div>
-              </div>
-              <button
-                type="button"
-                onClick={closePause}
-                aria-label="關閉"
-                className="text-petal-muted hover:text-petal-ink text-3xl leading-none p-2 -m-2"
-              >
-                ×
-              </button>
-            </div>
-
-            {/* Step 1 — Emotion selection */}
-            {pauseStep === 1 && (
-              <div className="flex-1 flex flex-col">
-                <h2 className="font-display text-2xl md:text-3xl font-light text-petal-ink leading-snug mb-2">
-                  你現在比較像哪一種狀態？
-                </h2>
-                <p className="font-body text-sm text-petal-ink-soft mb-8">
-                  選一個最接近的。這一步不是分析，是讓情緒被命名。
-                </p>
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-3 flex-1">
-                  {EMOTION_OPTIONS.map((opt) => (
-                    <button
-                      key={opt.key}
-                      type="button"
-                      onClick={() => pickEmotion(opt.key)}
-                      className="bg-white border-2 border-petal-rule rounded-md p-5 text-left hover:border-amber-400 hover:bg-amber-50/50 transition-colors min-h-[5.5rem]"
-                    >
-                      <div className="flex items-center gap-3">
-                        <span className="text-3xl leading-none" aria-hidden>{opt.emoji}</span>
-                        <div>
-                          <div className="font-display text-lg font-medium text-petal-ink">{opt.label}</div>
-                          <div className="font-body text-xs text-petal-ink-soft mt-0.5">{opt.sub}</div>
-                        </div>
-                      </div>
-                    </button>
-                  ))}
-                </div>
-              </div>
-            )}
-
-            {/* Step 2 — Safety phrase */}
-            {pauseStep === 2 && selectedPhrase && (
-              <div className="flex-1 flex flex-col">
-                <div className="font-body text-xs uppercase tracking-[0.14em] text-petal-muted mb-3">
-                  你選的狀態：{selectedPhrase.emoji} {selectedPhrase.label}
-                </div>
-                <h2 className="font-display text-2xl md:text-3xl font-light text-petal-ink leading-snug mb-2">
-                  照念這句話給對方聽
-                </h2>
-                <p className="font-body text-sm text-petal-ink-soft mb-8">
-                  不用自己想，直接念。重點是把界線講清楚，又不切斷連結。
-                </p>
-
-                <div className="bg-white border-2 border-amber-300 rounded-md p-6 md:p-8 mb-8 flex-1 flex items-center">
-                  <p className="font-display text-xl md:text-2xl font-light text-petal-ink leading-relaxed text-center w-full">
-                    「{selectedPhrase.phrase}」
-                  </p>
-                </div>
-
-                <div className="flex flex-col sm:flex-row gap-3">
-                  <button
-                    type="button"
-                    onClick={() => setPauseStep(1)}
-                    className="px-5 py-3 border border-petal-rule text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink rounded-md font-body text-sm transition-colors"
-                  >
-                    ← 換一個情緒
-                  </button>
-                  <button
-                    type="button"
-                    onClick={enterTurnTaking}
-                    className="flex-1 px-5 py-3 bg-amber-600 text-white rounded-md text-base font-medium hover:bg-amber-700 transition-colors"
-                  >
-                    我念完了，開始輪流說 →
-                  </button>
-                </div>
-              </div>
-            )}
-
-            {/* Step 3 — Enforced turn-taking */}
-            {pauseStep === 3 && (
-              <div className="flex-1 flex flex-col">
-                <div className="font-body text-xs uppercase tracking-[0.14em] text-petal-muted mb-2">
-                  第 {pauseRound} 輪 / 共 2 輪
-                </div>
-                <h2 className="font-display text-2xl md:text-3xl font-light text-petal-ink leading-snug mb-1">
-                  現在輪到：<span className="text-amber-700">{pauseRound === 1 ? 'A 方' : 'B 方'}</span> 說話
-                </h2>
-                <p className="font-body text-sm text-petal-ink-soft mb-6">
-                  {pauseRound === 1
-                    ? 'A 方＝剛剛選擇情緒的那位。另一方只能聽，不能打斷。'
-                    : 'B 方＝另一位。同樣只能聽，不能打斷。'}
-                </p>
-
-                {/* Timer */}
-                <div className="bg-white border-2 border-amber-200 rounded-md p-6 mb-6 text-center">
-                  <div className="font-body text-xs uppercase tracking-[0.14em] text-petal-muted mb-2">剩餘時間</div>
-                  <div className="font-display text-5xl md:text-6xl font-light text-amber-700 tabular-nums tracking-wide">
-                    {formatTime(pauseSeconds)}
-                  </div>
-                  <div className="w-full h-1.5 bg-amber-100 rounded-full mt-4 overflow-hidden">
-                    <div
-                      className="h-full bg-amber-500 transition-all duration-1000 ease-linear"
-                      style={{ width: `${(pauseSeconds / TURN_SECONDS) * 100}%` }}
-                    />
-                  </div>
-                </div>
-
-                {/* Listener reactions */}
-                <div className="flex-1">
-                  <p className="font-body text-xs uppercase tracking-[0.14em] text-petal-muted mb-3 text-center">
-                    聆聽者只能回應 — 不能打斷
-                  </p>
-                  <div className="grid grid-cols-2 gap-3">
-                    <button
-                      type="button"
-                      onClick={() => handleListenerReact('我在聽')}
-                      className="bg-white border-2 border-petal-rule rounded-md p-4 hover:border-amber-400 hover:bg-amber-50/40 transition-colors text-petal-ink font-body text-base"
-                    >
-                      <span className="text-2xl mr-2" aria-hidden>👍</span>
-                      我在聽
-                    </button>
-                    <button
-                      type="button"
-                      onClick={() => handleListenerReact('我理解你在努力說')}
-                      className="bg-white border-2 border-petal-rule rounded-md p-4 hover:border-amber-400 hover:bg-amber-50/40 transition-colors text-petal-ink font-body text-base"
-                    >
-                      <span className="text-2xl mr-2" aria-hidden>💛</span>
-                      我理解你在努力說
-                    </button>
-                  </div>
-                  {listenerNudge && (
-                    <p className="mt-4 font-display italic text-center text-amber-700 text-sm" role="status">
-                      ✓ {listenerNudge}
-                    </p>
-                  )}
-                </div>
-
-                {/* Skip turn */}
-                <div className="mt-6 pt-4 border-t border-petal-rule-soft text-center">
-                  <button
-                    type="button"
-                    onClick={skipTurn}
-                    className="font-body text-xs text-petal-muted hover:text-petal-ink transition-colors"
-                  >
-                    {pauseRound === 1 ? '已經講完，換對方 →' : '已經講完，進入收尾 →'}
-                  </button>
-                </div>
-              </div>
-            )}
-
-            {/* Step 4 — Closing */}
-            {pauseStep === 4 && (
-              <div className="flex-1 flex flex-col justify-center">
-                <h2 className="font-display text-2xl md:text-3xl font-light text-petal-ink leading-snug mb-2 text-center">
-                  你們做到了。
-                </h2>
-                <p className="font-body text-sm text-petal-ink-soft text-center mb-10 leading-relaxed">
-                  不一定要現在解決所有事，<br />
-                  能夠暫停、聽到彼此，已經是進展。
-                </p>
-
-                <div className="space-y-3">
-                  <button
-                    type="button"
-                    onClick={closePause}
-                    className="w-full px-5 py-4 bg-amber-600 text-white rounded-md text-base font-medium hover:bg-amber-700 transition-colors flex items-center justify-center gap-2"
-                  >
-                    <span className="text-xl" aria-hidden>🤝</span>
-                    我理解你現在的感覺
-                  </button>
-                  <button
-                    type="button"
-                    onClick={closePause}
-                    className="w-full px-5 py-4 bg-white border-2 border-amber-300 text-amber-900 rounded-md text-base font-medium hover:bg-amber-50 transition-colors flex items-center justify-center gap-2"
-                  >
-                    <Pause className="w-4 h-4 fill-amber-700 text-amber-700" strokeWidth={1.5} />
-                    我們先暫停 10 分鐘
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-    </div>
-    );
-  };
 
   // RoleplayView component moved to separate file
-
-  // Authentication Modal Component
-  const AuthModal = () => {
-    const [authMode, setAuthMode] = useState<'login' | 'register' | 'partner'>('login');
-    const [email, setEmail] = useState('');
-    const [nickname, setNickname] = useState('');
-    const [password, setPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
-    const [partnerCode, setPartnerCode] = useState('');
-    const [isLoading, setIsLoading] = useState(false);
-    useScrollLock(true);
-
-    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-      e.preventDefault();
-      
-      if (isLoading) return; // Prevent double-clicking
-      
-      setIsLoading(true);
-      
-      const formData = new FormData(e.currentTarget);
-      const emailValue = (formData.get('email') as string || '').trim();
-      const passwordValue = (formData.get('password') as string || '');
-      const nicknameValue = (formData.get('nickname') as string || '').trim();
-      const confirmPasswordValue = (formData.get('confirm-password') as string || '');
-      const partnerCodeValue = (formData.get('partner-code') as string || '').trim();
-      
-      try {
-        if (authMode === 'login') {
-          // Validate login form
-          if (!emailValue) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入電子郵件地址',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          if (!passwordValue) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入密碼',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          // Basic email validation
-          const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-          if (!emailRegex.test(emailValue)) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入有效的電子郵件地址',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          // Password length validation
-          if (passwordValue.length < 6) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '密碼至少需要6個字符',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          await handleLogin(emailValue, passwordValue);
-        } else if (authMode === 'register') {
-          // Validate registration form
-          if (!emailValue) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入電子郵件地址',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          if (!nicknameValue) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入暱稱',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          if (!passwordValue) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入密碼',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          // Basic email validation
-          const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-          if (!emailRegex.test(emailValue)) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入有效的電子郵件地址',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          // Nickname length validation
-          if (nicknameValue.length < 2 || nicknameValue.length > 50) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '暱稱必須在2-50個字符之間',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          // Password length validation
-          if (passwordValue.length < 6) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '密碼至少需要6個字符',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          if (passwordValue !== confirmPasswordValue) {
-            showNotification({
-              type: 'error',
-              title: '密碼不匹配',
-              message: '請確認兩次輸入的密碼相同',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          await handleRegister(emailValue, nicknameValue, passwordValue);
-        } else {
-          // Validate partner code
-          if (!partnerCodeValue) {
-            showNotification({
-              type: 'error',
-              title: '驗證錯誤',
-              message: '請輸入配對碼',
-              duration: 6000
-            });
-            setIsLoading(false);
-            return;
-          }
-          await handlePartnerConnect(partnerCodeValue);
-        }
-      } catch (error) {
-        // Error handling is done in individual functions
-        console.error('Form submission error:', error);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    return (
-      <div className="fixed inset-0 bg-petal-ink/40 backdrop-blur-sm flex items-center justify-center z-50 p-4">
-        <div className="bg-petal-cream rounded-md p-7 max-w-md w-full shadow-petal border border-petal-rule" data-testid="auth-modal">
-          <div className="flex justify-between items-end mb-6 pb-5 border-b border-petal-rule">
-            <div>
-              <div className="font-body text-[11px] font-medium uppercase tracking-[0.16em] text-petal-muted mb-2">
-                — {authMode === 'login' ? '登入' : authMode === 'register' ? '註冊' : '連接'}
-              </div>
-              <h3 data-testid="auth-modal-heading" className="font-display text-2xl font-light tracking-tight text-petal-ink">
-                {authMode === 'login' ? <>登入<em className="not-italic font-light italic text-pink-600">愛的時光</em></> :
-                 authMode === 'register' ? <>註冊<em className="not-italic font-light italic text-pink-600">新帳號</em></> :
-                 <>連接<em className="not-italic font-light italic text-pink-600">伴侶</em></>}
-              </h3>
-            </div>
-            <button
-              onClick={() => setShowAuthModal(false)}
-              data-testid="auth-modal-close-button"
-              className="text-petal-muted hover:text-petal-ink transition-colors"
-            >
-              <X className="w-5 h-5" strokeWidth={1.5} />
-            </button>
-          </div>
-
-          {authMode === 'login' && (
-            <form key="login-form" onSubmit={handleSubmit} className="space-y-4">
-              <div>
-                <label htmlFor="login-email" className="block text-sm font-medium text-gray-700 mb-2">
-                  電子郵件
-                </label>
-                <input
-                  id="login-email"
-                  name="email"
-                  data-testid="auth-email-input"
-                  type="email"
-                  defaultValue={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="輸入你的電子郵件"
-                  disabled={isLoading}
-                  autoComplete="username"
-                  required
-                />
-              </div>
-              <div>
-                <label htmlFor="login-password" className="block text-sm font-medium text-gray-700 mb-2">
-                  密碼
-                </label>
-                <input
-                  id="login-password"
-                  name="password"
-                  data-testid="auth-password-input"
-                  type="password"
-                  autoComplete="current-password"
-                  defaultValue={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="輸入你的密碼"
-                  disabled={isLoading}
-                  required
-                />
-              </div>
-              <button
-                type="submit"
-                disabled={isLoading}
-                data-testid="auth-submit-button"
-                className={`w-full py-3 rounded-md font-display italic text-base transition-colors ${
-                  isLoading
-                    ? 'bg-petal-cream-2 text-petal-muted cursor-not-allowed'
-                    : 'bg-petal-ink text-petal-cream hover:bg-pink-700'
-                }`}
-              >
-                {isLoading ? (
-                  <div className="flex items-center justify-center space-x-3">
-                    <div className="w-4 h-4 border-2 border-petal-muted border-t-transparent rounded-full animate-spin"></div>
-                    <span>登入中…</span>
-                  </div>
-                ) : (
-                  <span>開始愛的旅程 →</span>
-                )}
-              </button>
-            </form>
-          )}
-
-          {authMode === 'register' && (
-            <form key="register-form" onSubmit={handleSubmit} className="space-y-4">
-              <div>
-                <label htmlFor="register-email" className="block text-sm font-medium text-gray-700 mb-2">
-                  電子郵件
-                </label>
-                <input
-                  id="register-email"
-                  name="email"
-                  type="email"
-                  defaultValue={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="輸入你的電子郵件"
-                  disabled={isLoading}
-                  autoComplete="username"
-                  required
-                />
-              </div>
-              <div>
-                <label htmlFor="register-nickname" className="block text-sm font-medium text-gray-700 mb-2">
-                  暱稱
-                </label>
-                <input
-                  id="register-nickname"
-                  name="nickname"
-                  type="text"
-                  defaultValue={nickname}
-                  onChange={(e) => setNickname(e.target.value)}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="輸入你的暱稱"
-                  disabled={isLoading}
-                  autoComplete="nickname"
-                  required
-                />
-              </div>
-              <div>
-                <label htmlFor="register-password" className="block text-sm font-medium text-gray-700 mb-2">
-                  密碼
-                </label>
-                <input
-                  id="register-password"
-                  name="password"
-                  type="password"
-                  autoComplete="new-password"
-                  defaultValue={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="輸入你的密碼"
-                  disabled={isLoading}
-                  required
-                />
-              </div>
-              <div>
-                <label htmlFor="register-confirm-password" className="block text-sm font-medium text-gray-700 mb-2">
-                  確認密碼
-                </label>
-                <input
-                  id="register-confirm-password"
-                  name="confirm-password"
-                  type="password"
-                  autoComplete="new-password"
-                  defaultValue={confirmPassword}
-                  onChange={(e) => setConfirmPassword(e.target.value)}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="再次輸入密碼"
-                  disabled={isLoading}
-                  required
-                />
-              </div>
-              <button
-                type="submit"
-                disabled={isLoading}
-                data-testid="auth-submit-button"
-                className={`w-full py-3 rounded-md font-display italic text-base transition-colors ${
-                  isLoading
-                    ? 'bg-petal-cream-2 text-petal-muted cursor-not-allowed'
-                    : 'bg-petal-ink text-petal-cream hover:bg-pink-700'
-                }`}
-              >
-                {isLoading ? (
-                  <div className="flex items-center justify-center space-x-3">
-                    <div className="w-4 h-4 border-2 border-petal-muted border-t-transparent rounded-full animate-spin"></div>
-                    <span>註冊中…</span>
-                  </div>
-                ) : (
-                  <span>註冊帳號 →</span>
-                )}
-              </button>
-            </form>
-          )}
-
-          {authMode === 'partner' && (
-            <form key="partner-form" onSubmit={handleSubmit} className="space-y-4">
-              <div>
-                <label htmlFor="partner-code" className="block text-sm font-medium text-gray-700 mb-2">
-                  伴侶配對碼
-                </label>
-                <input
-                  id="partner-code"
-                  name="partner-code"
-                  type="text"
-                  defaultValue={partnerCode}
-                  onChange={(e) => setPartnerCode(e.target.value.toUpperCase())}
-                  className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-pink-500"
-                  placeholder="輸入伴侶的配對碼"
-                  disabled={isLoading}
-                  required
-                />
-              </div>
-              <button
-                type="submit"
-                disabled={isLoading}
-                data-testid="auth-submit-button"
-                className={`w-full py-3 rounded-md font-display italic text-base transition-colors ${
-                  isLoading
-                    ? 'bg-petal-cream-2 text-petal-muted cursor-not-allowed'
-                    : 'bg-petal-ink text-petal-cream hover:bg-pink-700'
-                }`}
-              >
-                {isLoading ? (
-                  <div className="flex items-center justify-center space-x-3">
-                    <div className="w-4 h-4 border-2 border-petal-muted border-t-transparent rounded-full animate-spin"></div>
-                    <span>連接中…</span>
-                  </div>
-                ) : (
-                  <span>連接伴侶 →</span>
-                )}
-              </button>
-            </form>
-          )}
-
-          <div className="mt-4 text-center space-y-2">
-            {authMode === 'login' && (
-              <>
-                <button
-                  onClick={() => setAuthMode('register')}
-                  disabled={isLoading}
-                  className={`text-pink-600 hover:text-pink-700 text-sm block w-full ${
-                    isLoading ? 'opacity-50 cursor-not-allowed' : ''
-                  }`}
-                >
-                  還沒帳號？立即註冊
-                </button>
-                <button
-                  onClick={() => setAuthMode('partner')}
-                  disabled={isLoading}
-                  className={`text-pink-600 hover:text-pink-700 text-sm block w-full ${
-                    isLoading ? 'opacity-50 cursor-not-allowed' : ''
-                  }`}
-                >
-                  已有帳號？連接伴侶
-                </button>
-              </>
-            )}
-            {authMode === 'register' && (
-              <button
-                onClick={() => setAuthMode('login')}
-                disabled={isLoading}
-                className={`text-pink-600 hover:text-pink-700 text-sm block w-full ${
-                  isLoading ? 'opacity-50 cursor-not-allowed' : ''
-                }`}
-              >
-                已有帳號？立即登入
-              </button>
-            )}
-            {authMode === 'partner' && (
-              <button
-                onClick={() => setAuthMode('login')}
-                disabled={isLoading}
-                className={`text-pink-600 hover:text-pink-700 text-sm block w-full ${
-                  isLoading ? 'opacity-50 cursor-not-allowed' : ''
-                }`}
-              >
-                還沒帳號？立即註冊
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-  // Coin Shop Component
-  const CoinShopView = () => {
-    const [showAddGiftModal, setShowAddGiftModal] = useState(false);
-    useScrollLock(showAddGiftModal);
-    const [newGift, setNewGift] = useState({
-      title: '',
-      description: '',
-      cost: 1000,
-      category: 'service' as CoinGift['category'],
-      icon: '🎁'
-    });
-
-    const allGifts = [...defaultGifts, ...customGifts];
-
-    const handleAddGift = (e: React.FormEvent) => {
-      e.preventDefault();
-      addCustomGift(newGift.title, newGift.description, newGift.cost, newGift.category, newGift.icon);
-      setNewGift({ title: '', description: '', cost: 1000, category: 'service', icon: '🎁' });
-    };
-
-    return (
-      <div className="space-y-10">
-        <div className="border-b border-petal-rule pb-7">
-          <div className="flex flex-col md:flex-row md:items-end md:justify-between gap-6">
-            <div>
-              <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-                — 商店
-              </div>
-              <h2 className="font-display text-4xl md:text-5xl font-light tracking-tight text-petal-ink leading-[1.05] mb-3">
-                金幣<em className="not-italic font-light italic text-pink-600">商店</em>
-              </h2>
-              <p className="font-display italic font-light text-base text-petal-muted">
-                用愛賺來的金幣，兌換特別禮品。
-              </p>
-            </div>
-            <div className="flex flex-col items-start md:items-end gap-2">
-              <div className="font-display italic font-light text-2xl text-petal-ink">
-                <Coins className="inline w-4 h-4 mr-1.5 text-petal-rose-deep" strokeWidth={1.5} />
-                <b className="not-italic font-medium">{totalCoins}</b> <span className="text-base text-petal-muted">枚</span>
-              </div>
-              <button
-                onClick={() => setShowAddGiftModal(true)}
-                className="px-4 py-1.5 border border-petal-rule rounded-full text-xs font-body text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink transition-colors flex items-center space-x-1.5"
-              >
-                <Plus className="w-3.5 h-3.5" />
-                <span>自訂禮品</span>
-              </button>
-            </div>
-          </div>
-        </div>
-
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5">
-          {allGifts.map((gift) => (
-            <div key={gift.id} className="bg-white rounded-md border border-petal-rule p-6 hover:border-petal-rose transition-colors">
-              <div className="mb-4 pb-4 border-b border-petal-rule-soft">
-                <div className="text-2xl mb-3 opacity-75 saturate-75">{gift.icon}</div>
-                <h3 className="font-display text-lg font-medium tracking-tight text-petal-ink mb-1">{gift.title}</h3>
-                <p className="font-body text-sm text-petal-ink-soft leading-relaxed">{gift.description}</p>
-              </div>
-
-              <div className="flex items-center justify-between mb-4">
-                <span className="font-body text-[11px] uppercase tracking-[0.12em] text-petal-muted">
-                  {gift.category === 'service' ? '服務' :
-                   gift.category === 'experience' ? '體驗' :
-                   gift.category === 'physical' ? '實物' : '親密'}
-                </span>
-                <div className="font-display italic font-light text-base text-petal-ink">
-                  <Coins className="inline w-3.5 h-3.5 mr-1 text-petal-rose-deep" strokeWidth={1.5} />
-                  <b className="not-italic font-normal">{gift.cost}</b>
-                </div>
-              </div>
-
-              <button
-                onClick={() => purchaseGift(gift)}
-                disabled={totalCoins < gift.cost}
-                className={`w-full py-2.5 rounded-md font-display italic text-base transition-colors ${
-                  totalCoins >= gift.cost
-                    ? 'bg-petal-ink text-petal-cream hover:bg-pink-700'
-                    : 'bg-petal-cream-2 text-petal-muted cursor-not-allowed'
-                }`}
-              >
-                {totalCoins >= gift.cost ? '立即兌換 →' : '金幣不足'}
-              </button>
-            </div>
-          ))}
-        </div>
-
-        {/* Add Custom Gift Modal */}
-        {showAddGiftModal && (
-          <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
-            <div className="bg-white rounded-lg p-6 max-w-md w-full">
-              <div className="flex justify-between items-center mb-4">
-                <h3 className="text-xl font-bold text-yellow-700">添加自訂禮品</h3>
-                <button
-                  onClick={() => setShowAddGiftModal(false)}
-                  className="text-gray-500 hover:text-gray-700"
-                >
-                  <X className="w-6 h-6" />
-                </button>
-              </div>
-
-              <form onSubmit={handleAddGift} className="space-y-4">
-                <div>
-                  <label htmlFor="gift-title" className="block text-sm font-medium text-gray-700 mb-2">
-                    禮品名稱
-                  </label>
-                  <input
-                    id="gift-title"
-                    name="gift-title"
-                    type="text"
-                    value={newGift.title}
-                    onChange={(e) => setNewGift(prev => ({ ...prev, title: e.target.value }))}
-                    className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-yellow-500"
-                    placeholder="例如：按摩服務"
-                    required
-                  />
-                </div>
-
-                <div>
-                  <label htmlFor="gift-description" className="block text-sm font-medium text-gray-700 mb-2">
-                    描述
-                  </label>
-                  <textarea
-                    id="gift-description"
-                    name="gift-description"
-                    value={newGift.description}
-                    onChange={(e) => setNewGift(prev => ({ ...prev, description: e.target.value }))}
-                    className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-yellow-500"
-                    placeholder="詳細描述這個禮品"
-                    rows={3}
-                    required
-                  />
-                </div>
-
-                <div className="grid grid-cols-2 gap-4">
-                  <div>
-                    <label htmlFor="gift-cost" className="block text-sm font-medium text-gray-700 mb-2">
-                      金幣價格
-                    </label>
-                    <input
-                      id="gift-cost"
-                      name="gift-cost"
-                      type="number"
-                      value={newGift.cost}
-                      onChange={(e) => setNewGift(prev => ({ ...prev, cost: parseInt(e.target.value) }))}
-                      className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-yellow-500"
-                      min="100"
-                      step="100"
-                      required
-                    />
-                  </div>
-
-                  <div>
-                    <label htmlFor="gift-icon" className="block text-sm font-medium text-gray-700 mb-2">
-                      圖示
-                    </label>
-                    <input
-                      id="gift-icon"
-                      name="gift-icon"
-                      type="text"
-                      value={newGift.icon}
-                      onChange={(e) => setNewGift(prev => ({ ...prev, icon: e.target.value }))}
-                      className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-yellow-500"
-                      placeholder="🎁"
-                      required
-                    />
-                  </div>
-                </div>
-
-                <div>
-                  <label htmlFor="gift-category" className="block text-sm font-medium text-gray-700 mb-2">
-                    類別
-                  </label>
-                  <select
-                    id="gift-category"
-                    name="gift-category"
-                    value={newGift.category}
-                    onChange={(e) => setNewGift(prev => ({ ...prev, category: e.target.value as CoinGift['category'] }))}
-                    className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-yellow-500"
-                    required
-                  >
-                    <option value="service">服務</option>
-                    <option value="experience">體驗</option>
-                    <option value="physical">實物</option>
-                    <option value="intimate">親密</option>
-                  </select>
-                </div>
-
-                <button
-                  type="submit"
-                  className="w-full bg-petal-ink text-petal-cream py-3 rounded-md font-display italic text-base hover:bg-pink-700 transition-colors"
-                >
-                  添加禮品
-                </button>
-              </form>
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  // Foreplay View Component
-  const ForeplayView = () => {
-    const [selectedActivity, setSelectedActivity] = useState<ForeplayActivity | null>(null);
-    const [selectedPosition, setSelectedPosition] = useState<PositionSuggestion | null>(null);
-    useScrollLock(!!selectedActivity);
-    useScrollLock(!!selectedPosition);
-
-    const handleTryActivity = async (activity: ForeplayActivity) => {
-      const coinsEarned = activity.coins;
-
-      // Update coins via backend API
-      try {
-        await apiService.updateCoins(coinsEarned);
-        setTotalCoins(prev => prev + coinsEarned);
-      } catch (error) {
-        console.warn('Failed to update coins via API, using local update only:', error);
-        setTotalCoins(prev => prev + coinsEarned);
-      }
-
-      showNotification({
-        type: 'success',
-        title: `已嘗試 ${activity.title}！`,
-        message: '記得稍後記錄你們的親密時光',
-        coins: coinsEarned,
-        duration: 4000
-      });
-    };
-
-    const handleTryPosition = async (position: PositionSuggestion) => {
-      const coinsEarned = position.coins;
-
-      // Update coins via backend API
-      try {
-        await apiService.updateCoins(coinsEarned);
-        setTotalCoins(prev => prev + coinsEarned);
-      } catch (error) {
-        console.warn('Failed to update coins via API, using local update only:', error);
-        setTotalCoins(prev => prev + coinsEarned);
-      }
-
-      showNotification({
-        type: 'success',
-        title: `已嘗試 ${position.name}！`,
-        message: '記得稍後記錄你們的親密時光',
-        coins: coinsEarned,
-        duration: 4000
-      });
-    };
-
-    const handleCompleteCombo = async (combo: typeof comboSuggestions[number]) => {
-      const coinsEarned = combo.bonusCoins;
-
-      try {
-        await apiService.updateCoins(coinsEarned);
-        setTotalCoins(prev => prev + coinsEarned);
-      } catch (error) {
-        console.warn('Failed to update coins via API, using local update only:', error);
-        setTotalCoins(prev => prev + coinsEarned);
-      }
-
-      showNotification({
-        type: 'success',
-        title: `組合技達成 — ${combo.name}！`,
-        message: `連續完成 ${combo.sequence.length} 種姿勢，記得記錄這次親密時光`,
-        coins: coinsEarned,
-        duration: 5000,
-      });
-    };
-
-    return (
-      <div className="space-y-10">
-        {/* Foreplay Activities */}
-        <div id="foreplay-activities" className="scroll-mt-24">
-          <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink mb-6">
-            前戲<em className="not-italic font-light italic text-pink-600">活動</em>
-          </h3>
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
-            {foreplayActivities.map((activity, index) => (
-              <div key={index} className="border border-petal-rule rounded-md p-5 bg-white hover:border-petal-rose transition-colors">
-                <div className="flex justify-between items-start mb-3 pb-3 border-b border-petal-rule-soft">
-                  <h4 className="font-display text-lg font-medium tracking-tight text-petal-ink">{activity.title}</h4>
-                  <div className="font-display italic font-light text-sm text-petal-rose-deep whitespace-nowrap">
-                    <Coins className="inline w-3.5 h-3.5 mr-1" strokeWidth={1.5} />
-                    +{activity.coins}
-                  </div>
-                </div>
-                <p className="font-body text-sm text-petal-ink-soft leading-relaxed mb-4">{activity.description}</p>
-                <div className="flex items-center justify-between mb-3">
-                  <span className="font-body text-xs text-petal-muted flex items-center">
-                    <Clock className="w-3.5 h-3.5 mr-1" strokeWidth={1.5} />
-                    {activity.duration}
-                  </span>
-                  <button
-                    onClick={() => handleTryActivity(activity)}
-                    className="px-4 py-1.5 bg-petal-ink text-petal-cream rounded-full hover:bg-pink-700 transition-colors flex items-center space-x-1.5 font-display italic text-sm"
-                  >
-                    <Play className="w-3.5 h-3.5" strokeWidth={1.5} />
-                    <span>嘗試</span>
-                  </button>
-                </div>
-                <button
-                  onClick={() => setSelectedActivity(activity)}
-                  className="font-body text-xs text-petal-ink-soft hover:text-petal-rose-deep transition-colors"
-                >
-                  查看詳細提示 →
-                </button>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {/* Position Suggestions */}
-        <div id="position-suggestions" className="scroll-mt-24">
-          <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink mb-6">
-            姿勢<em className="not-italic font-light italic text-pink-600">建議</em>
-          </h3>
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4">
-            {positionSuggestions.map((position, index) => (
-              <div key={index} className="border border-petal-rule rounded-md p-5 bg-white hover:border-petal-rose transition-colors">
-                <div className="flex justify-between items-start mb-3">
-                  <h4 className="font-display text-base font-medium tracking-tight text-petal-ink">{position.name}</h4>
-                  <div className="font-display italic font-light text-xs text-petal-rose-deep whitespace-nowrap">
-                    <Coins className="inline w-3 h-3 mr-0.5" strokeWidth={1.5} />
-                    +{position.coins}
-                  </div>
-                </div>
-                <div className="mb-3">
-                  <span className={`inline-block px-2.5 py-0.5 rounded-full text-[10px] border ${
-                    position.difficulty === '簡單' ? 'border-petal-sage/60 bg-petal-sage/10 text-petal-sage-deep' :
-                    position.difficulty === '中等' ? 'border-petal-rose-soft bg-petal-rose-soft/30 text-petal-rose-deep' :
-                    'border-petal-ink-soft/30 bg-petal-ink-soft/5 text-petal-ink-soft'
-                  }`}>
-                    {position.difficulty}
-                  </span>
-                </div>
-                <p className="font-body text-sm text-petal-ink-soft leading-relaxed mb-3">{position.description}</p>
-                <div className="flex justify-between items-center">
-                  <button
-                    onClick={() => setSelectedPosition(position)}
-                    className="font-body text-xs text-petal-ink-soft hover:text-petal-rose-deep transition-colors"
-                  >
-                    詳細資訊
-                  </button>
-                  <button
-                    onClick={() => handleTryPosition(position)}
-                    className="bg-petal-ink text-petal-cream px-3 py-1 rounded-full font-body text-xs hover:bg-pink-700 transition-colors"
-                  >
-                    嘗試
-                  </button>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {/* Combo Suggestions — 連續組合技 */}
-        <div id="combo-suggestions" className="scroll-mt-24">
-          <div className="flex items-baseline justify-between mb-6">
-            <h3 className="font-display text-2xl font-medium tracking-tight text-petal-ink">
-              連續<em className="not-italic font-light italic text-pink-600">組合技</em>
-            </h3>
-            <span className="font-display italic font-light text-sm text-petal-muted">
-              一次連續多種姿勢，額外金幣獎勵
-            </span>
-          </div>
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-            {comboSuggestions.map((combo, index) => (
-              <div key={index} className="border border-petal-rose-soft bg-petal-rose-soft/10 rounded-md p-5 hover:border-petal-rose transition-colors flex flex-col">
-                <div className="flex justify-between items-start mb-3">
-                  <h4 className="font-display text-lg font-medium tracking-tight text-petal-ink">{combo.name}</h4>
-                  <span className={`inline-block px-2.5 py-0.5 rounded-full text-[10px] border ${
-                    combo.difficulty === '簡單' ? 'border-petal-sage/60 bg-petal-sage/10 text-petal-sage-deep' :
-                    combo.difficulty === '中等' ? 'border-petal-rose-soft bg-petal-rose-soft/30 text-petal-rose-deep' :
-                    'border-petal-ink-soft/30 bg-petal-ink-soft/5 text-petal-ink-soft'
-                  }`}>
-                    {combo.difficulty}
-                  </span>
-                </div>
-                <p className="font-body text-sm text-petal-ink-soft leading-relaxed mb-4">{combo.description}</p>
-                <ol className="space-y-1.5 mb-4">
-                  {combo.sequence.map((step, i) => (
-                    <li key={i} className="flex items-center font-body text-sm text-petal-ink">
-                      <span className="w-5 h-5 rounded-full bg-petal-ink text-petal-cream font-display italic text-[11px] flex items-center justify-center mr-2.5 flex-shrink-0">
-                        {i + 1}
-                      </span>
-                      {step}
-                    </li>
-                  ))}
-                </ol>
-                <div className="mt-auto flex items-center justify-between pt-3 border-t border-petal-rule-soft">
-                  <div className="font-display italic font-light text-sm text-petal-rose-deep whitespace-nowrap">
-                    <Coins className="inline w-3.5 h-3.5 mr-1" strokeWidth={1.5} />
-                    +{combo.bonusCoins} 獎勵
-                  </div>
-                  <button
-                    onClick={() => handleCompleteCombo(combo)}
-                    className="px-4 py-1.5 bg-petal-ink text-petal-cream rounded-full hover:bg-pink-700 transition-colors flex items-center space-x-1.5 font-display italic text-sm"
-                  >
-                    <Sparkles className="w-3.5 h-3.5" strokeWidth={1.5} />
-                    <span>完成組合技</span>
-                  </button>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {/* Activity Detail Modal */}
-        {selectedActivity && (
-          <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
-            <div className="bg-white rounded-lg p-6 max-w-md w-full">
-              <div className="flex justify-between items-start mb-4">
-                <h3 className="text-xl font-bold text-pink-700">{selectedActivity.title}</h3>
-                <button
-                  onClick={() => setSelectedActivity(null)}
-                  className="text-gray-500 hover:text-gray-700"
-                >
-                  <X className="w-6 h-6" />
-                </button>
-              </div>
-              <p className="text-gray-600 mb-4">{selectedActivity.description}</p>
-              <div className="mb-4">
-                <h4 className="font-semibold text-gray-800 mb-2">實用提示：</h4>
-                <ul className="space-y-1">
-                  {selectedActivity.tips.map((tip: string, index: number) => (
-                    <li key={index} className="text-sm text-gray-600 flex items-start">
-                      <span className="text-pink-500 mr-2">•</span>
-                      {tip}
-                    </li>
-                  ))}
-                </ul>
-              </div>
-              <button
-                onClick={() => {
-                  handleTryActivity(selectedActivity);
-                  setSelectedActivity(null);
-                }}
-                className="w-full bg-petal-ink text-petal-cream py-3 rounded-md font-display italic text-base hover:bg-pink-700 transition-colors"
-              >
-                開始嘗試
-              </button>
-            </div>
-          </div>
-        )}
-
-        {/* Position Detail Modal */}
-        {selectedPosition && (
-          <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
-            <div className="bg-white rounded-lg p-6 max-w-md w-full">
-              <div className="flex justify-between items-start mb-4">
-                <h3 className="text-xl font-bold text-pink-700">{selectedPosition.name}</h3>
-                <button
-                  onClick={() => setSelectedPosition(null)}
-                  className="text-gray-500 hover:text-gray-700"
-                >
-                  <X className="w-6 h-6" />
-                </button>
-              </div>
-              <p className="text-gray-600 mb-4">{selectedPosition.description}</p>
-              <div className="mb-4">
-                <h4 className="font-semibold text-gray-800 mb-2">優點：</h4>
-                <ul className="space-y-1">
-                  {selectedPosition.benefits.map((benefit: string, index: number) => (
-                    <li key={index} className="text-sm text-gray-600 flex items-start">
-                      <span className="text-pink-500 mr-2">•</span>
-                      {benefit}
-                    </li>
-                  ))}
-                </ul>
-              </div>
-              <button
-                onClick={() => {
-                  handleTryPosition(selectedPosition);
-                  setSelectedPosition(null);
-                }}
-                className="w-full bg-petal-ink text-petal-cream py-3 rounded-md font-display italic text-base hover:bg-pink-700 transition-colors"
-              >
-                嘗試這個姿勢
-              </button>
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  };
 
   const handleNicknameChange = useCallback((partner: 'partner1' | 'partner2', value: string) => {
     setNicknames(prev => ({...prev, [partner]: value}));
   }, []);
 
   // SettingsView component moved to separate file
-
-  // Our Journey View
-  const OurJourneyView = () => {
-    const sortedMilestones = [...journeyMilestones]
-      .filter(m => m.date || m.place)
-      .sort((a, b) => {
-        const da = a.date ? new Date(a.date).getTime() : 0;
-        const db = b.date ? new Date(b.date).getTime() : 0;
-        return da - db;
-      });
-
-    const handleMilestoneClick = (milestone: JourneyMilestone) => {
-      if (milestone.recordId) {
-        // Find and highlight the specific record
-        const record = intimateRecords.find(r => r.id === milestone.recordId);
-        if (record) {
-          setCurrentView('calendar');
-          // Could add more specific navigation logic here
-        }
-      }
-    };
-
-    return (
-      <div className="space-y-10">
-        <div className="border-b border-petal-rule pb-7">
-          <div className="font-body text-[11px] font-medium uppercase tracking-[0.18em] text-petal-muted mb-3">
-            — 旅程
-          </div>
-          <h2 className="font-display text-4xl md:text-5xl font-light tracking-tight text-petal-ink leading-[1.05] mb-3">
-            我們的<em className="not-italic font-light italic text-pink-600">愛情旅程</em>
-          </h2>
-          <p className="font-display italic font-light text-base text-petal-muted">
-            記錄每個重要的時刻和里程碑。
-          </p>
-        </div>
-
-        <div>
-          <div className="relative">
-            {/* Timeline Line */}
-            <div className="absolute left-6 top-0 bottom-0 w-px bg-petal-rule"></div>
-
-            <div className="space-y-8">
-              {sortedMilestones.map((milestone) => (
-                <div key={milestone.id} className="relative flex items-start space-x-6">
-                  {/* Timeline Node */}
-                  <div className="w-12 h-12 rounded-full bg-petal-cream border border-petal-rose-soft flex items-center justify-center relative z-10 text-base opacity-80 saturate-75">
-                    {milestone.type === 'meeting' ? '💕' :
-                     milestone.type === 'first_date' ? '🌹' :
-                     milestone.type === 'first_kiss' ? '💋' :
-                     milestone.type === 'first_sex' ? '💋' :
-                     milestone.type === 'marriage' ? '👫' :
-                     milestone.type === 'child_born' ? '👶' :
-                     '✦'}
-                  </div>
-
-                  {/* Content */}
-                  <div className="flex-1 bg-white border border-petal-rule rounded-md p-5">
-                    <div className="flex items-start justify-between mb-2">
-                      <div>
-                        <h3 className="font-display text-lg font-medium tracking-tight text-petal-ink">{milestone.title}</h3>
-                        <p className="font-display italic font-light text-sm text-petal-muted mt-0.5">
-                          {milestone.date ? milestone.date.slice(0, 10) : (milestone.place ? '—' : '')}
-                        </p>
-                        {milestone.place && (
-                          <p className="font-body text-xs text-petal-muted">地點：{milestone.place}</p>
-                        )}
-                      </div>
-                      {milestone.count && (
-                        <span className="font-display italic font-light text-xs text-petal-rose-deep border border-petal-rose-soft px-2.5 py-0.5 rounded-full">
-                          第 {milestone.count} 次
-                        </span>
-                      )}
-                    </div>
-                    <p className="font-body text-sm text-petal-ink-soft leading-relaxed mb-3">{milestone.description}</p>
-
-                    {milestone.recordId && (
-                      <button
-                        onClick={() => handleMilestoneClick(milestone)}
-                        className="inline-flex items-center font-body text-xs text-petal-ink-soft hover:text-petal-rose-deep transition-colors"
-                      >
-                        <Heart className="w-3 h-3 mr-1" strokeWidth={1.5} />
-                        查看詳細記錄 →
-                      </button>
-                    )}
-                  </div>
-                </div>
-              ))}
-            </div>
-
-            {/* Future Milestones Preview */}
-            <div className="mt-8 p-6 bg-petal-cream-2/40 rounded-md border border-dashed border-petal-rose-soft">
-              <h3 className="font-display text-lg font-medium tracking-tight text-petal-ink mb-4 flex items-center">
-                <Sparkles className="w-4 h-4 mr-2 text-petal-rose-deep" strokeWidth={1.5} />
-                即將到來的里程碑
-              </h3>
-              <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                {[
-                  { count: 10, achieved: intimateRecords.length >= 10 },
-                  { count: 20, achieved: intimateRecords.length >= 20 },
-                  { count: 50, achieved: intimateRecords.length >= 50 },
-                  { count: 100, achieved: intimateRecords.length >= 100 }
-                ].map(({ count, achieved }) => (
-                  <div key={count} className={`p-4 rounded-lg border-2 ${
-                    achieved 
-                      ? 'border-green-200 bg-green-50' 
-                      : 'border-gray-200 bg-white'
-                  }`}>
-                    <div className="flex items-center justify-between">
-                      <span className="font-medium text-gray-800">第 {count} 次親密時光</span>
-                      {achieved ? (
-                        <span className="text-green-600 font-bold">✓ 已達成</span>
-                      ) : (
-                        <span className="text-gray-500">
-                          還需 {count - intimateRecords.length} 次
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  };
 
   const navItems = [
     { id: 'record', label: '記錄時光', icon: Calendar },
@@ -5280,12 +1818,61 @@ const LoveTimeApp = () => {
     switch (currentView) {
       case 'record':
       case 'achievements':
-        return <CalendarView />;
-      case 'shop': return <CoinShopView />;
+        return (
+          <CalendarView
+            selectedDate={selectedDate}
+            setSelectedDate={setSelectedDate}
+            intimateRecords={intimateRecords}
+            setIntimateRecords={setIntimateRecords}
+            cycleRecords={cycleRecords}
+            setCycleRecords={setCycleRecords}
+            authState={authState}
+            calendarMonth={calendarMonth}
+            setCalendarMonth={setCalendarMonth}
+            editingRecord={editingRecord}
+            setEditingRecord={setEditingRecord}
+            showRecordModal={showRecordModal}
+            setShowRecordModal={setShowRecordModal}
+            setSelectedRecord={setSelectedRecord}
+            setShowRecordDetail={setShowRecordDetail}
+            setDayPickerDate={setDayPickerDate}
+            setDayPickerRecords={setDayPickerRecords}
+            setCurrentView={setCurrentView}
+            addIntimateRecord={addIntimateRecord}
+            showNotification={showNotification}
+            showRecordDetails={showRecordDetails}
+            openDeleteConfirm={openDeleteConfirm}
+            defaultRoleplayScripts={defaultRoleplayScripts}
+            customScripts={customScripts}
+            togetherSince={togetherSince}
+            daysTogether={daysTogether}
+            primaryTimezone={primaryTimezone}
+          />
+        );
+      case 'shop': return (
+        <CoinShopView
+          totalCoins={totalCoins}
+          defaultGifts={defaultGifts}
+          customGifts={customGifts}
+          addCustomGift={addCustomGift}
+          purchaseGift={purchaseGift}
+        />
+      );
       case 'foreplay':
       case 'games':
-        return <GamesView />;
-      case 'conflict': return <ConflictView />;
+        return <GamesView
+        totalCoins={totalCoins}
+        customMemoryQuestions={customMemoryQuestions}
+        customEmotions={customEmotions}
+        setTotalCoins={setTotalCoins}
+        showNotification={showNotification}
+      />;
+      case 'conflict': return (
+        <ConflictView
+          showNotification={showNotification}
+          partnerConnected={partnerConnected}
+        />
+      );
       case 'events': return (
         <EventsView
           authState={authState}
@@ -5311,9 +1898,21 @@ const LoveTimeApp = () => {
         onToggleFavorite={toggleFavoriteScript}
         initialScriptTitle={pendingScriptTitle}
         onInitialScriptConsumed={() => setPendingScriptTitle(null)}
-        renderGames={() => <GamesView />}
+        renderGames={() => <GamesView
+        totalCoins={totalCoins}
+        customMemoryQuestions={customMemoryQuestions}
+        customEmotions={customEmotions}
+        setTotalCoins={setTotalCoins}
+        showNotification={showNotification}
+      />}
       />;
-      case 'journey': return <OurJourneyView />;
+      case 'journey': return (
+        <OurJourneyView
+          journeyMilestones={journeyMilestones}
+          intimateRecords={intimateRecords}
+          setCurrentView={setCurrentView}
+        />
+      );
       case 'wall': return <WallView
         authState={authState}
         nicknames={nicknames}
@@ -5344,7 +1943,13 @@ const LoveTimeApp = () => {
         />
       );
       case 'upgrade': return <UpgradeView reason={upgradeReason} showNotification={showNotification} />;
-      default: return <GamesView />;
+      default: return <GamesView
+        totalCoins={totalCoins}
+        customMemoryQuestions={customMemoryQuestions}
+        customEmotions={customEmotions}
+        setTotalCoins={setTotalCoins}
+        showNotification={showNotification}
+      />;
     }
   };
 
@@ -5352,118 +1957,6 @@ const LoveTimeApp = () => {
     setNotifications(prev => prev.filter(n => n.id !== id));
   };
 
-
-  // Calendar component for date picking
-  const CalendarDatePicker = ({ selectedDate, onDateSelect }: { selectedDate: string, onDateSelect: (date: string) => void }) => {
-    const [currentMonth, setCurrentMonth] = useState(() => {
-      const date = selectedDate ? new Date(selectedDate) : new Date();
-      return new Date(date.getFullYear(), date.getMonth(), 1);
-    });
-
-    const getDaysInMonth = (date: Date) => {
-      const year = date.getFullYear();
-      const month = date.getMonth();
-      const firstDay = new Date(year, month, 1);
-      const lastDay = new Date(year, month + 1, 0);
-      const daysInMonth = lastDay.getDate();
-      const startingDayOfWeek = firstDay.getDay();
-
-      const days = [];
-      
-      // Previous month's trailing days
-      for (let i = startingDayOfWeek - 1; i >= 0; i--) {
-        const day = new Date(year, month, -i);
-        days.push({ date: day, isCurrentMonth: false });
-      }
-      
-      // Current month's days
-      for (let day = 1; day <= daysInMonth; day++) {
-        const date = new Date(year, month, day);
-        days.push({ date, isCurrentMonth: true });
-      }
-      
-      // Next month's leading days
-      const remainingDays = 42 - days.length; // 6 rows × 7 days
-      for (let day = 1; day <= remainingDays; day++) {
-        const date = new Date(year, month + 1, day);
-        days.push({ date, isCurrentMonth: false });
-      }
-      
-      return days;
-    };
-
-    const formatDate = (date: Date) => {
-      // Grid cells are local-time Dates from `new Date(y, m, d)` — read local
-      // components directly. toISOString() would shift to UTC and mis-key.
-      const y = date.getFullYear();
-      const m = String(date.getMonth() + 1).padStart(2, '0');
-      const d = String(date.getDate()).padStart(2, '0');
-      return `${y}-${m}-${d}`;
-    };
-
-    const isSelected = (date: Date) => {
-      return formatDate(date) === selectedDate;
-    };
-
-    const isToday = (date: Date) => {
-      const today = new Date();
-      return date.toDateString() === today.toDateString();
-    };
-
-    const days = getDaysInMonth(currentMonth);
-    const monthYear = formatMonthYear(currentMonth, primaryTimezone);
-
-    return (
-      <div className="bg-white rounded-md p-5 border border-petal-rule">
-        <div className="flex items-center justify-between mb-4">
-          <button
-            onClick={() => setCurrentMonth(new Date(currentMonth.getFullYear(), currentMonth.getMonth() - 1, 1))}
-            className="w-8 h-8 border border-petal-rule rounded-full text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink transition-colors"
-          >
-            ‹
-          </button>
-          <h3 className="font-display italic font-light text-lg text-petal-ink">{monthYear}</h3>
-          <button
-            onClick={() => setCurrentMonth(new Date(currentMonth.getFullYear(), currentMonth.getMonth() + 1, 1))}
-            className="w-8 h-8 border border-petal-rule rounded-full text-petal-ink-soft hover:border-petal-ink hover:text-petal-ink transition-colors"
-          >
-            ›
-          </button>
-        </div>
-
-        <div className="grid grid-cols-7 gap-1 mb-2">
-          {['日', '一', '二', '三', '四', '五', '六'].map(day => (
-            <div key={day} className="p-2 text-center font-body text-[10px] font-medium uppercase tracking-[0.14em] text-petal-muted">
-              {day}
-            </div>
-          ))}
-        </div>
-        
-        <div className="grid grid-cols-7 gap-1">
-          {days.map((dayInfo, index) => {
-            const { date, isCurrentMonth } = dayInfo;
-            const selected = isSelected(date);
-            const today = isToday(date);
-
-            return (
-              <button
-                key={index}
-                onClick={() => onDateSelect(formatDate(date))}
-                className={`
-                  p-2 font-display text-sm rounded-full hover:bg-petal-cream-2 transition-colors
-                  ${isCurrentMonth ? 'text-petal-ink' : 'text-petal-rule'}
-                  ${selected ? 'bg-petal-rose-deep text-petal-cream hover:bg-petal-rose-deep italic' : ''}
-                  ${today && !selected ? 'border border-petal-rose-deep text-petal-rose-deep' : ''}
-                `}
-              >
-                {date.getDate()}
-              </button>
-            );
-          })}
-        </div>
-      </div>
-    );
-  };
 
   const primaryTimezone = getPrimaryTimezone({
     couplePrimaryTz: authState.user?.couplePrimaryTimezone,
@@ -5668,7 +2161,15 @@ const LoveTimeApp = () => {
       </div>
       
       {/* Modals */}
-      {showAuthModal && <AuthModal />}
+      {showAuthModal && (
+        <AuthModal
+          onClose={() => setShowAuthModal(false)}
+          showNotification={showNotification}
+          handleLogin={handleLogin}
+          handleRegister={handleRegister}
+          handlePartnerConnect={handlePartnerConnect}
+        />
+      )}
       {showScriptUploadModal && (
         <ScriptUploadModal
           editingScript={editingScript}
